refactor(layout): decompose mod.rs into 7 focused modules
Extract ~3,700 lines from the monolithic 5,012-line mod.rs into:
- extraction.rs: JSON-to-typed data conversion, tick helpers
- note_layout.rs: note positioning, glyph generation
- annotations.rs: ties, slurs, dots, ledger lines
- barlines.rs: barline rendering and segments
- staff_groups.rs: multi-staff collision, brackets
- structural.rs: clefs, key/time signature glyphs
- assembly.rs: staff lines, system assembly

mod.rs is now a thin orchestrator (1,323 lines including tests).
All 160 tests pass, clippy clean, WASM builds successfully.
Public API surface unchanged.

Closes #52
