+++ backend/src/layout/mod.rs
@@ -5,7 +5,6 @@
 //! virtualization boundary and provides bounding boxes for efficient rendering,
 //! hit testing, and interaction.
 
-use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
 pub mod batcher;
@@ -17,129 +16,30 @@ pub mod spacer;
 pub mod stems;
 pub mod types;
 
+pub(crate) mod annotations;
+pub(crate) mod assembly;
+pub(crate) mod barlines;
+pub(crate) mod extraction;
+pub(crate) mod note_layout;
+pub(crate) mod staff_groups;
+pub(crate) mod structural;
+
 #[cfg(target_arch = "wasm32")]
 pub mod wasm;
 
 pub use breaker::MeasureInfo;
+pub use extraction::NoteData;
 pub use types::{
     BarLine, BarLineSegment, BarLineType, BoundingBox, BracketGlyph, BracketType, Color,
-    GlobalLayout, Glyph, GlyphRun, LedgerLine, MeasureNumber, NameLabel, Point, RepeatDotPosition,
-    SourceReference, Staff, StaffGroup, StaffLine, System, TickRange, VoltaBracketLayout,
+    GlobalLayout, Glyph, GlyphRun, LayoutConfig, LedgerLine, MeasureNumber, NameLabel, Point,
+    RepeatDotPosition, SourceReference, Staff, StaffGroup, StaffLine, System, TickRange,
+    VoltaBracketLayout,
 };
 
-/// Configuration for layout computation
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct LayoutConfig {
-    /// Maximum system width in logical units (default: 1600.0)
-    pub max_system_width: f32,
-    /// Scaling factor: logical units per staff space (default: 10.0)
-    pub units_per_space: f32,
-    /// Vertical spacing between systems in logical units (default: 150.0)
-    pub system_spacing: f32,
-    /// System height in logical units (default: 200.0 for grand staff)
-    pub system_height: f32,
-}
-
-impl Default for LayoutConfig {
-    fn default() -> Self {
-        Self {
-            max_system_width: 2400.0, // Wide enough for 3+ measures per system
-            units_per_space: 20.0,    // SMuFL: font_size 80 = 4 spaces, so 1 space = 20 units
-            system_spacing: 100.0,    // Spacing between systems (gap after system_height)
-            system_height: 200.0,     // Base height for a single staff system
-        }
-    }
-}
-
-/// Compute the start tick of a measure, accounting for pickup/anacrusis.
-fn measure_start_tick(measure_index: usize, pickup_ticks: u32, ticks_per_measure: u32) -> u32 {
-    if pickup_ticks > 0 {
-        if measure_index == 0 {
-            0
-        } else {
-            pickup_ticks + (measure_index as u32 - 1) * ticks_per_measure
-        }
-    } else {
-        measure_index as u32 * ticks_per_measure
-    }
-}
-
-/// Compute the end tick of a measure, accounting for pickup/anacrusis.
-fn measure_end_tick(measure_index: usize, pickup_ticks: u32, ticks_per_measure: u32) -> u32 {
-    if pickup_ticks > 0 {
-        if measure_index == 0 {
-            pickup_ticks
-        } else {
-            pickup_ticks + measure_index as u32 * ticks_per_measure
-        }
-    } else {
-        (measure_index as u32 + 1) * ticks_per_measure
-    }
-}
-
-/// Map a tick position to its measure index, accounting for pickup/anacrusis.
-fn tick_to_measure_index(tick: u32, pickup_ticks: u32, ticks_per_measure: u32) -> usize {
-    if pickup_ticks > 0 {
-        if tick < pickup_ticks {
-            0
-        } else {
-            ((tick - pickup_ticks) / ticks_per_measure) as usize + 1
-        }
-    } else {
-        (tick / ticks_per_measure) as usize
-    }
-}
-
-/// Measure start tick using actual boundaries if available, formula fallback.
-fn actual_start(
-    measure_index: usize,
-    measure_end_ticks: &[u32],
-    pickup_ticks: u32,
-    ticks_per_measure: u32,
-) -> u32 {
-    if !measure_end_ticks.is_empty()
-        && measure_index > 0
-        && measure_index <= measure_end_ticks.len()
-    {
-        measure_end_ticks[measure_index - 1]
-    } else if measure_index == 0 {
-        0
-    } else {
-        measure_start_tick(measure_index, pickup_ticks, ticks_per_measure)
-    }
-}
-
-/// Measure end tick using actual boundaries if available, formula fallback.
-fn actual_end(
-    measure_index: usize,
-    measure_end_ticks: &[u32],
-    pickup_ticks: u32,
-    ticks_per_measure: u32,
-) -> u32 {
-    if measure_index < measure_end_ticks.len() {
-        measure_end_ticks[measure_index]
-    } else {
-        measure_end_tick(measure_index, pickup_ticks, ticks_per_measure)
-    }
-}
-
-/// Map tick to measure index using actual boundaries, with formula fallback.
-fn actual_tick_to_measure(
-    tick: u32,
-    measure_end_ticks: &[u32],
-    pickup_ticks: u32,
-    ticks_per_measure: u32,
-) -> usize {
-    if !measure_end_ticks.is_empty() {
-        // Binary search: find the first measure whose end_tick > tick
-        match measure_end_ticks.binary_search(&(tick + 1)) {
-            Ok(i) => i,
-            Err(i) => i.min(measure_end_ticks.len().saturating_sub(1)),
-        }
-    } else {
-        tick_to_measure_index(tick, pickup_ticks, ticks_per_measure)
-    }
-}
+use extraction::{
+    RestLayoutEvent, StaffData, actual_end, actual_start, actual_tick_to_measure,
+    extract_instruments, extract_measures,
+};
 
 /// Compute layout from a CompiledScore
 ///
@@ -406,7 +306,7 @@ pub fn compute_layout(score: &serde_json::Value, config: &LayoutConfig) -> Globa
             .filter(|(t, _)| *t > system.tick_range.start_tick && *t < system.tick_range.end_tick)
             .map(|(t, _)| *t)
             .collect();
-        let note_positions = compute_unified_note_positions(
+        let note_positions = note_layout::compute_unified_note_positions(
             &all_staves,
             &system.tick_range,
             system.bounding_box.width,
@@ -496,7 +396,13 @@ pub fn compute_layout(score: &serde_json::Value, config: &LayoutConfig) -> Globa
         let staff_extents: Vec<(f32, f32)> = instruments
             .iter()
             .flat_map(|inst| inst.staves.iter())
-            .map(|sd| compute_staff_note_extents(sd, &system.tick_range, config.units_per_space))
+            .map(|sd| {
+                note_layout::compute_staff_note_extents(
+                    sd,
+                    &system.tick_range,
+                    config.units_per_space,
+                )
+            })
             .collect();
 
         // For each staff, record whether an inter-instrument gap precedes it
@@ -507,25 +413,14 @@ pub fn compute_layout(score: &serde_json::Value, config: &LayoutConfig) -> Globa
             }
         }
 
-        // Compute cumulative collision-avoidance extra per staff
-        let min_clearance = 2.0 * config.units_per_space; // 2 staff-space clearance (accounts for stems/accidentals)
-        let mut cumulative_collision_extra: Vec<f32> = vec![0.0; staff_extents.len()];
-        for i in 0..staff_extents.len().saturating_sub(1) {
-            let (_, max_y_upper) = staff_extents[i]; // bottom extent of upper staff
-            let (min_y_lower, _) = staff_extents[i + 1]; // top extent of lower staff
-
-            // Default spacing between origins of staff i and staff i+1
-            let mut pair_spacing = intra_staff_multiplier * config.units_per_space;
-            if has_inter_gap_before[i + 1] {
-                pair_spacing += inter_instrument_multiplier * config.units_per_space;
-            }
-
-            // Needed spacing so bottom of upper staff is min_clearance above top of lower
-            let needed = max_y_upper - min_y_lower + min_clearance;
-            let extra = (needed - pair_spacing).max(0.0);
-            cumulative_collision_extra[i + 1] = cumulative_collision_extra[i] + extra;
-        }
-        let total_collision_extra = cumulative_collision_extra.last().copied().unwrap_or(0.0);
+        let (cumulative_collision_extra, total_collision_extra) =
+            staff_groups::compute_collision_gaps(
+                &staff_extents,
+                &has_inter_gap_before,
+                intra_staff_multiplier,
+                inter_instrument_multiplier,
+                config.units_per_space,
+            );
 
         for (instrument_index, instrument) in instruments.iter().enumerate() {
             let mut staves = Vec::new();
@@ -548,7 +443,7 @@ pub fn compute_layout(score: &serde_json::Value, config: &LayoutConfig) -> Globa
                     + cumulative_collision_extra[absolute_staff_index];
 
                 // Position glyphs for this staff using unified note positions
-                let glyphs = position_glyphs_for_staff(
+                let glyphs = note_layout::position_glyphs_for_staff(
                     staff_data,
                     &system.tick_range,
                     config.units_per_space,
@@ -575,163 +470,26 @@ pub fn compute_layout(score: &serde_json::Value, config: &LayoutConfig) -> Globa
                 glyph_runs.extend(batcher::batch_glyphs(pseudo_glyphs));
 
                 // Create staff lines (5 lines evenly spaced)
-                let staff_lines = create_staff_lines(
+                let staff_lines = assembly::create_staff_lines(
                     staff_vertical_offset,
                     system.bounding_box.width,
                     config.units_per_space,
                 );
 
                 // T036-T037: Generate structural glyphs (clef, time sig, key sig) at system start
-                let mut structural_glyphs = Vec::new();
-
-                // Determine the active key signature at this system's start tick
-                let system_key_sharps = staff_data.get_key_at_tick(system.tick_range.start_tick);
-
-                // Determine the active clef at this system's start tick
-                let system_clef = staff_data.get_clef_at_tick(system.tick_range.start_tick);
-
-                // Position clef at x=60 (left margin with room for brace and glyph extent)
-                let clef_glyph = positioner::position_clef(
-                    system_clef,
-                    60.0,
-                    config.units_per_space,
+                let structural_glyphs = structural::render_structural_glyphs(
+                    staff_data,
+                    &system.tick_range,
+                    system.index,
+                    system.bounding_box.width,
                     staff_vertical_offset,
-                );
-                structural_glyphs.push(clef_glyph);
-
-                // Position key signature after clef (x=120)
-                let key_sig_glyphs = positioner::position_key_signature(
-                    system_key_sharps,
-                    system_clef,
-                    120.0,
                     config.units_per_space,
-                    staff_vertical_offset,
+                    &note_positions,
+                    &measure_x_bounds,
                 );
-                structural_glyphs.extend(key_sig_glyphs);
-
-                // Position time signature after key signature — only on the first system
-                // (standard engraving: repeat only when the time signature changes)
-                if system.index == 0 {
-                    let key_sig_width = system_key_sharps.abs() as f32 * 15.0;
-                    let time_sig_x = 120.0 + key_sig_width + 20.0; // Add 20 unit gap
-                    let time_sig_glyphs = positioner::position_time_signature(
-                        staff_data.time_numerator,
-                        staff_data.time_denominator,
-                        time_sig_x,
-                        config.units_per_space,
-                        staff_vertical_offset,
-                    );
-                    structural_glyphs.extend(time_sig_glyphs);
-                }
-
-                // Render key signature changes at mid-system measure boundaries
-                if !staff_data.key_signature_events.is_empty() {
-                    for &(event_tick, event_sharps) in &staff_data.key_signature_events {
-                        // Skip if event is outside this system's tick range
-                        if event_tick <= system.tick_range.start_tick
-                            || event_tick >= system.tick_range.end_tick
-                        {
-                            continue;
-                        }
-                        // Look up the x position for this tick from measure bounds
-                        if let Some(&(measure_x_start, _)) = measure_x_bounds.get(&event_tick) {
-                            // Position the key signature glyphs at the measure start
-                            // Shift slightly right to leave room after the barline
-                            let key_x = measure_x_start + 10.0;
-                            let active_clef = staff_data.get_clef_at_tick(event_tick);
-                            let mid_key_glyphs = positioner::position_key_signature(
-                                event_sharps,
-                                active_clef,
-                                key_x,
-                                config.units_per_space,
-                                staff_vertical_offset,
-                            );
-                            structural_glyphs.extend(mid_key_glyphs);
-                        }
-                    }
-                }
-
-                // Render clef changes within this system (mid-system and
-                // mid-measure).  Also render a courtesy clef at the right
-                // edge when the clef changes at the start of the next system.
-                if !staff_data.clef_events.is_empty() {
-                    for (event_tick, event_clef) in &staff_data.clef_events {
-                        // Skip the initial clef (tick 0 or system start) — it is
-                        // already rendered as the system-start clef glyph.
-                        if *event_tick <= system.tick_range.start_tick {
-                            continue;
-                        }
-
-                        // Courtesy clef: event falls on or after this system's
-                        // end tick → place a small warning clef at the right edge.
-                        if *event_tick >= system.tick_range.end_tick {
-                            if *event_tick == system.tick_range.end_tick {
-                                let courtesy_x = system.bounding_box.width - 40.0;
-                                let courtesy_glyph = positioner::position_courtesy_clef(
-                                    event_clef,
-                                    courtesy_x,
-                                    config.units_per_space,
-                                    staff_vertical_offset,
-                                );
-                                structural_glyphs.push(courtesy_glyph);
-                            }
-                            continue;
-                        }
-
-                        // Event is inside this system.  Find the enclosing
-                        // measure: the one whose start_tick is the largest value
-                        // that is ≤ event_tick.
-                        let enclosing = measure_x_bounds
-                            .iter()
-                            .filter(|(tick, _)| **tick <= *event_tick)
-                            .max_by_key(|(tick, _)| *tick);
-
-                        if let Some((&_m_tick, &(measure_x_start, _measure_x_end))) = enclosing {
-                            // Place the courtesy clef near the start of the
-                            // measure (for measure-start changes) or just before
-                            // the first note at the clef change tick for
-                            // mid-measure changes.
-                            let clef_x = if *event_tick == _m_tick {
-                                measure_x_start + 10.0
-                            } else {
-                                // Mid-measure: use the pre-computed note position
-                                // at the clef change tick and place the clef just
-                                // before it.  Fall back to the measure start if no
-                                // note exists at the exact tick.
-                                if let Some(&note_x) = note_positions.get(event_tick) {
-                                    // Place clef well before the note (extra 50 units
-                                    // reserved in spacing + clef glyph width ~30)
-                                    note_x - 45.0
-                                } else {
-                                    // No note at this exact tick — find the nearest
-                                    // note after the event tick within this system.
-                                    let next_note_x = note_positions
-                                        .iter()
-                                        .filter(|(t, _)| {
-                                            **t > *event_tick && **t < system.tick_range.end_tick
-                                        })
-                                        .min_by_key(|(t, _)| *t)
-                                        .map(|(_, &x)| x);
-                                    if let Some(nx) = next_note_x {
-                                        nx - 45.0
-                                    } else {
-                                        measure_x_start + 10.0
-                                    }
-                                }
-                            };
-                            let mid_clef_glyph = positioner::position_courtesy_clef(
-                                event_clef,
-                                clef_x,
-                                config.units_per_space,
-                                staff_vertical_offset,
-                            );
-                            structural_glyphs.push(mid_clef_glyph);
-                        }
-                    }
-                }
 
                 // Create bar lines at measure boundaries
-                let bar_lines = create_bar_lines(
+                let bar_lines = barlines::create_bar_lines(
                     &measure_infos,
                     &system.tick_range,
                     staff_vertical_offset,
@@ -742,664 +500,17 @@ pub fn compute_layout(score: &serde_json::Value, config: &LayoutConfig) -> Globa
                     &measure_x_bounds,
                 );
 
-                // Generate ledger lines for notes outside the 5-line staff
-                let mut ledger_lines = Vec::new();
-                for voice in &staff_data.voices {
-                    let notes_in_range: Vec<NoteData> = voice
-                        .notes
-                        .iter()
-                        .filter(|note| {
-                            note.start_tick >= system.tick_range.start_tick
-                                && note.start_tick < system.tick_range.end_tick
-                        })
-                        .map(|note| {
-                            (
-                                note.pitch,
-                                note.start_tick,
-                                note.duration_ticks,
-                                note.spelling,
-                                note.staccato,
-                                note.dot_count,
-                            )
-                        })
-                        .collect();
-                    let offsets: Vec<f32> = notes_in_range
-                        .iter()
-                        .map(|(_, tick, _, _, _, _)| *note_positions.get(tick).unwrap_or(&0.0))
-                        .collect();
-                    let ledger_clefs: Vec<&str> = notes_in_range
-                        .iter()
-                        .map(|(_, tick, _, _, _, _)| staff_data.get_clef_at_tick(*tick))
-                        .collect();
-                    ledger_lines.extend(positioner::position_ledger_lines(
-                        &notes_in_range,
-                        &offsets,
-                        &ledger_clefs,
-                        config.units_per_space,
-                        staff_vertical_offset,
-                    ));
-                }
-
-                // Compute notation dots (augmentation dots and staccato dots)
-                let mut notation_dots = Vec::new();
-                let dot_radius = 0.18 * config.units_per_space;
-                let staff_middle_y = staff_vertical_offset + 1.5 * config.units_per_space;
-                let num_voices_for_dots = staff_data.voices.len();
-                for (dot_voice_idx, voice) in staff_data.voices.iter().enumerate() {
-                    // Multi-voice stem rule for dot placement
-                    let forced_stem_down_dots: Option<bool> = if num_voices_for_dots > 1 {
-                        Some(dot_voice_idx > 0)
-                    } else {
-                        None
-                    };
-                    let voice_notes: Vec<&NoteEvent> = voice
-                        .notes
-                        .iter()
-                        .filter(|n| {
-                            n.start_tick >= system.tick_range.start_tick
-                                && n.start_tick < system.tick_range.end_tick
-                        })
-                        .collect();
-                    // Group notes by tick for chord handling
-                    let mut tick_groups: std::collections::BTreeMap<u32, Vec<&NoteEvent>> =
-                        std::collections::BTreeMap::new();
-                    for note in &voice_notes {
-                        tick_groups.entry(note.start_tick).or_default().push(note);
-                    }
-
-                    // Pre-compute beam group stem directions so staccato dots on beamed
-                    // notes use the actual beam group direction, not per-chord heuristics.
-                    let beam_tick_stem_down: std::collections::HashMap<u32, bool> = {
-                        let mut map = std::collections::HashMap::new();
-                        if forced_stem_down_dots.is_none() {
-                            // Build beamable notes for beam group analysis
-                            let beamable_raw: Vec<beams::BeamableNote> = voice_notes
-                                .iter()
-                                .filter(|n| n.duration_ticks <= 480 || !n.beam_info.is_empty())
-                                .map(|n| {
-                                    let clef = staff_data.get_clef_at_tick(n.start_tick);
-                                    let y = positioner::pitch_to_y_with_spelling(
-                                        n.pitch,
-                                        clef,
-                                        config.units_per_space,
-                                        n.spelling,
-                                    ) + staff_vertical_offset;
-                                    let beam_types: Vec<String> =
-                                        n.beam_info.iter().map(|(_, bt)| bt.clone()).collect();
-                                    beams::BeamableNote {
-                                        x: *note_positions.get(&n.start_tick).unwrap_or(&0.0),
-                                        y,
-                                        stem_end_y: 0.0,
-                                        tick: n.start_tick,
-                                        duration_ticks: n.duration_ticks,
-                                        beam_levels: n.beam_info.len() as u8,
-                                        beam_types,
-                                        event_index: 0,
-                                    }
-                                })
-                                .collect();
-                            // Deduplicate by tick (chords share a stem)
-                            let mut deduped: Vec<beams::BeamableNote> = Vec::new();
-                            let mut seen: std::collections::HashSet<u32> =
-                                std::collections::HashSet::new();
-                            for n in beamable_raw {
-                                if seen.insert(n.tick) {
-                                    deduped.push(n);
-                                }
-                            }
-                            let has_beam_info = deduped.iter().any(|n| !n.beam_types.is_empty());
-                            let groups = if has_beam_info {
-                                beams::build_beam_groups_from_musicxml(&deduped)
-                            } else {
-                                let gs = beams::group_beamable_by_time_signature(
-                                    &deduped,
-                                    staff_data.time_numerator,
-                                    staff_data.time_denominator,
-                                );
-                                gs.into_iter()
-                                    .map(|notes| beams::BeamGroup {
-                                        beam_count: 1,
-                                        notes,
-                                    })
-                                    .collect()
-                            };
-                            for group in &groups {
-                                if group.notes.len() < 2 {
-                                    continue;
-                                }
-                                let dir = beams::compute_group_stem_direction(
-                                    &group.notes,
-                                    staff_middle_y,
-                                );
-                                let is_down = dir == stems::StemDirection::Down;
-                                for note in &group.notes {
-                                    map.insert(note.tick, is_down);
-                                }
-                            }
-                        }
-                        map
-                    };
-
-                    for (_tick, group) in &tick_groups {
-                        let clef = staff_data.get_clef_at_tick(*_tick);
-                        // Compute y positions for all notes in this chord/tick
-                        let mut note_ys: Vec<(f32, &NoteEvent)> = group
-                            .iter()
-                            .map(|n| {
-                                let y = positioner::pitch_to_y_with_spelling(
-                                    n.pitch,
-                                    clef,
-                                    config.units_per_space,
-                                    n.spelling,
-                                ) + staff_vertical_offset;
-                                (y, *n)
-                            })
-                            .collect();
-                        note_ys.sort_by(|a, b| a.0.partial_cmp(&b.0).unwrap());
-
-                        // Determine stem direction for the chord/note
-                        // Priority: multi-voice override > beam group direction > per-chord rule
-                        let stem_down = if let Some(forced) = forced_stem_down_dots {
-                            forced
-                        } else if let Some(&beamed_down) = beam_tick_stem_down.get(_tick) {
-                            beamed_down
-                        } else {
-                            let most_extreme_y = note_ys
-                                .iter()
-                                .max_by(|a, b| {
-                                    let da = (a.0 - staff_middle_y).abs();
-                                    let db = (b.0 - staff_middle_y).abs();
-                                    da.partial_cmp(&db).unwrap()
-                                })
-                                .map(|(y, _)| *y)
-                                .unwrap_or(staff_middle_y);
-                            most_extreme_y <= staff_middle_y
-                        };
-
-                        // Augmentation dots: placed to the right of each notehead
-                        for &(y_raw, note) in &note_ys {
-                            if note.dot_count > 0 {
-                                let note_x = *note_positions.get(&note.start_tick).unwrap_or(&0.0);
-                                // Shift y to visual center (pitch_to_y has -0.5*ups text offset)
-                                let visual_y = y_raw + 0.5 * config.units_per_space;
-                                // If note sits on a staff line, shift dot up to nearest space
-                                let dot_y = shift_dot_to_space(
-                                    visual_y,
-                                    staff_vertical_offset,
-                                    config.units_per_space,
-                                );
-                                let notehead_half_w = stems::Stem::NOTEHEAD_WIDTH;
-                                for d in 0..note.dot_count {
-                                    let gap = 0.4 * config.units_per_space;
-                                    let dot_spacing = 0.6 * config.units_per_space;
-                                    notation_dots.push(types::NotationDot {
-                                        x: note_x + notehead_half_w + gap + d as f32 * dot_spacing,
-                                        y: dot_y,
-                                        radius: dot_radius,
-                                    });
-                                }
-                            }
-                        }
-
-                        // Staccato dot: one per chord, placed on the notehead side (opposite stem)
-                        let has_staccato = group.iter().any(|n| n.staccato);
-                        if has_staccato {
-                            // Anchor on the notehead side (opposite the stem)
-                            let (anchor_y_raw, anchor_note) = if stem_down {
-                                // Stem down → staccato above the highest note (lowest y)
-                                note_ys[0]
-                            } else {
-                                // Stem up → staccato below the lowest note (highest y)
-                                *note_ys.last().unwrap()
-                            };
-                            let note_x =
-                                *note_positions.get(&anchor_note.start_tick).unwrap_or(&0.0);
-                            let visual_y = anchor_y_raw + 0.5 * config.units_per_space;
-                            let staccato_offset = 1.2 * config.units_per_space;
-                            let staccato_y = if stem_down {
-                                // Stem down → dot above notehead (negative direction)
-                                visual_y - staccato_offset
-                            } else {
-                                // Stem up → dot below notehead (positive direction)
-                                visual_y + staccato_offset
-                            };
-                            // Shift to space if on a line
-                            let staccato_y = shift_dot_to_space(
-                                staccato_y,
-                                staff_vertical_offset,
-                                config.units_per_space,
-                            );
-                            notation_dots.push(types::NotationDot {
-                                x: note_x,
-                                y: staccato_y,
-                                radius: dot_radius,
-                            });
-                        }
-                    }
-                }
-
-                // Compute tie arcs for tied notes
-                let mut tie_arcs: Vec<types::TieArc> = Vec::new();
-                let mut slur_arcs: Vec<types::TieArc> = Vec::new();
-                let notehead_half_w = stems::Stem::NOTEHEAD_WIDTH;
-                {
-                    // Build a lookup: note_id → (x, visual_y, pitch, start_tick)
-                    let mut note_lookup: std::collections::HashMap<&str, (f32, f32, u8, u32)> =
-                        std::collections::HashMap::new();
-
-                    for voice in &staff_data.voices {
-                        for n in &voice.notes {
-                            if n.note_id.is_empty() {
-                                continue;
-                            }
-                            let in_range = n.start_tick >= system.tick_range.start_tick
-                                && n.start_tick < system.tick_range.end_tick;
-                            if !in_range {
-                                continue;
-                            }
-                            let clef = staff_data.get_clef_at_tick(n.start_tick);
-                            let y_raw = positioner::pitch_to_y_with_spelling(
-                                n.pitch,
-                                clef,
-                                config.units_per_space,
-                                n.spelling,
-                            ) + staff_vertical_offset;
-                            let visual_y = y_raw + 0.5 * config.units_per_space;
-                            let note_x = *note_positions.get(&n.start_tick).unwrap_or(&0.0);
-                            note_lookup
-                                .insert(&n.note_id, (note_x, visual_y, n.pitch, n.start_tick));
-                        }
-                    }
-
-                    // Collect tied note pitches per tick for chord tie detection
-                    let mut tied_pitches_per_tick: std::collections::HashMap<u32, Vec<u8>> =
-                        std::collections::HashMap::new();
-                    for voice in &staff_data.voices {
-                        for n in &voice.notes {
-                            if n.tie_next.is_some() && !n.note_id.is_empty() {
-                                let in_range = n.start_tick >= system.tick_range.start_tick
-                                    && n.start_tick < system.tick_range.end_tick;
-                                if in_range {
-                                    tied_pitches_per_tick
-                                        .entry(n.start_tick)
-                                        .or_default()
-                                        .push(n.pitch);
-                                }
-                            }
-                        }
-                    }
-
-                    // Helper: determine tie arc direction (above/below notehead)
-                    let determine_tie_above =
-                        |pitch: u8,
-                         start_tick: u32,
-                         spelling: Option<(char, i8)>,
-                         tied_pitches: &std::collections::HashMap<u32, Vec<u8>>|
-                         -> bool {
-                            if let Some(pitches) = tied_pitches.get(&start_tick) {
-                                if pitches.len() > 1 {
-                                    let max_pitch = pitches.iter().copied().max().unwrap_or(0);
-                                    let min_pitch = pitches.iter().copied().min().unwrap_or(0);
-                                    if pitch == max_pitch {
-                                        return true;
-                                    } else if pitch == min_pitch {
-                                        return false;
-                                    }
-                                }
-                            }
-                            // Single tie or middle note: stem direction
-                            let clef = staff_data.get_clef_at_tick(start_tick);
-                            let y_raw = positioner::pitch_to_y_with_spelling(
-                                pitch,
-                                clef,
-                                config.units_per_space,
-                                spelling,
-                            ) + staff_vertical_offset;
-                            y_raw <= staff_middle_y
-                        };
-
-                    let system_right_edge = system.bounding_box.width;
-
-                    // For each note with tie_next, find the target and compute arc
-                    for voice in &staff_data.voices {
-                        for n in &voice.notes {
-                            let tie_target_id = match &n.tie_next {
-                                Some(id) => id.as_str(),
-                                None => continue,
-                            };
-                            if n.note_id.is_empty() {
-                                continue;
-                            }
-                            let in_range = n.start_tick >= system.tick_range.start_tick
-                                && n.start_tick < system.tick_range.end_tick;
-                            if !in_range {
-                                continue;
-                            }
-
-                            let start_info = match note_lookup.get(n.note_id.as_str()) {
-                                Some(info) => *info,
-                                None => continue,
-                            };
-
-                            let (start_x, start_y, _start_pitch, start_tick) = start_info;
-
-                            let above = determine_tie_above(
-                                n.pitch,
-                                start_tick,
-                                n.spelling,
-                                &tied_pitches_per_tick,
-                            );
-
-                            match note_lookup.get(tie_target_id) {
-                                Some(&(end_x, end_y, _end_pitch, _end_tick)) => {
-                                    // Same-system tie: full arc
-                                    let arc_start_x = start_x + notehead_half_w;
-                                    let arc_end_x = end_x - notehead_half_w;
-                                    let arc_start_y = start_y;
-                                    let arc_end_y = end_y;
-
-                                    let span_x = (arc_end_x - arc_start_x).abs();
-                                    let arc_height = span_x.mul_add(0.15, 0.0).clamp(4.0, 30.0);
-
-                                    let y_offset = if above { -arc_height } else { arc_height };
-                                    let mid_y = (arc_start_y + arc_end_y) / 2.0 + y_offset;
-
-                                    tie_arcs.push(types::TieArc {
-                                        start: types::Point {
-                                            x: arc_start_x,
-                                            y: arc_start_y,
-                                        },
-                                        end: types::Point {
-                                            x: arc_end_x,
-                                            y: arc_end_y,
-                                        },
-                                        cp1: types::Point {
-                                            x: arc_start_x + span_x * 0.2,
-                                            y: mid_y,
-                                        },
-                                        cp2: types::Point {
-                                            x: arc_end_x - span_x * 0.2,
-                                            y: mid_y,
-                                        },
-                                        above,
-                                        note_id_start: n.note_id.clone(),
-                                        note_id_end: tie_target_id.to_string(),
-                                    });
-                                }
-                                None => {
-                                    // Cross-system outgoing tie: arc from start note to system right edge
-                                    let arc_start_x = start_x + notehead_half_w;
-                                    let arc_end_x = system_right_edge;
-                                    let arc_y = start_y;
-
-                                    let span_x = arc_end_x - arc_start_x;
-                                    let arc_height = span_x.mul_add(0.15, 0.0).clamp(4.0, 30.0);
-                                    let y_offset = if above { -arc_height } else { arc_height };
-                                    let mid_y = arc_y + y_offset;
-
-                                    tie_arcs.push(types::TieArc {
-                                        start: types::Point {
-                                            x: arc_start_x,
-                                            y: arc_y,
-                                        },
-                                        end: types::Point {
-                                            x: arc_end_x,
-                                            y: arc_y,
-                                        },
-                                        cp1: types::Point {
-                                            x: arc_start_x + span_x * 0.2,
-                                            y: mid_y,
-                                        },
-                                        cp2: types::Point {
-                                            x: arc_end_x - span_x * 0.2,
-                                            y: mid_y,
-                                        },
-                                        above,
-                                        note_id_start: n.note_id.clone(),
-                                        note_id_end: tie_target_id.to_string(),
-                                    });
-                                }
-                            }
-                        }
-                    }
-
-                    // Cross-system incoming ties: notes in previous systems with
-                    // tie_next targeting a note in this system
-                    {
-                        // Build a set of note IDs in this system for quick lookup
-                        let system_note_ids: std::collections::HashSet<&str> =
-                            note_lookup.keys().copied().collect();
-
-                        for voice in &staff_data.voices {
-                            for n in &voice.notes {
-                                let tie_target_id = match &n.tie_next {
-                                    Some(id) => id.as_str(),
-                                    None => continue,
-                                };
-                                // Only consider notes BEFORE this system
-                                if n.start_tick >= system.tick_range.start_tick {
-                                    continue;
-                                }
-                                // Target must be in this system
-                                if !system_note_ids.contains(tie_target_id) {
-                                    continue;
-                                }
-                                let (end_x, end_y, _end_pitch, end_tick) =
-                                    *note_lookup.get(tie_target_id).unwrap();
-
-                                let above = determine_tie_above(
-                                    n.pitch,
-                                    end_tick,
-                                    n.spelling,
-                                    &tied_pitches_per_tick,
-                                );
-
-                                // Arc from left edge of system to target note
-                                let arc_end_x = end_x - notehead_half_w;
-                                // Start before note area; if target is the first note,
-                                // ensure the arc has a minimum visible span.
-                                let incoming_offset = 3.0 * notehead_half_w;
-                                let arc_start_x = (unified_left_margin - incoming_offset)
-                                    .min(arc_end_x - 20.0)
-                                    .max(0.0);
-                                let arc_y = end_y;
-
-                                let span_x = (arc_end_x - arc_start_x).max(1.0);
-                                let arc_height = span_x.mul_add(0.15, 0.0).clamp(4.0, 30.0);
-                                let y_offset = if above { -arc_height } else { arc_height };
-                                let mid_y = arc_y + y_offset;
-
-                                tie_arcs.push(types::TieArc {
-                                    start: types::Point {
-                                        x: arc_start_x,
-                                        y: arc_y,
-                                    },
-                                    end: types::Point {
-                                        x: arc_end_x,
-                                        y: arc_y,
-                                    },
-                                    cp1: types::Point {
-                                        x: arc_start_x + span_x * 0.2,
-                                        y: mid_y,
-                                    },
-                                    cp2: types::Point {
-                                        x: arc_end_x - span_x * 0.2,
-                                        y: mid_y,
-                                    },
-                                    above,
-                                    note_id_start: n.note_id.clone(),
-                                    note_id_end: tie_target_id.to_string(),
-                                });
-                            }
-                        }
-                    }
-
-                    // Compute slur arcs (phrase marks) — same 3-case logic as ties
-                    // Slur endpoints attach to notehead edge (top/bottom), not center.
-                    // X position is near notehead center for natural curve attachment.
-                    {
-                        let system_note_ids: std::collections::HashSet<&str> =
-                            note_lookup.keys().copied().collect();
-                        let notehead_half_h = config.units_per_space * 0.5;
-
-                        for voice in &staff_data.voices {
-                            for n in &voice.notes {
-                                let slur_target_id = match &n.slur_next {
-                                    Some(id) => id.as_str(),
-                                    None => continue,
-                                };
-                                if n.note_id.is_empty() {
-                                    continue;
-                                }
-
-                                // Slur direction: use MusicXML placement if available,
-                                // otherwise fall back to staff-position heuristic.
-                                let above = match n.slur_above {
-                                    Some(v) => v,
-                                    None => {
-                                        let clef = staff_data.get_clef_at_tick(n.start_tick);
-                                        let y_raw = positioner::pitch_to_y_with_spelling(
-                                            n.pitch,
-                                            clef,
-                                            config.units_per_space,
-                                            n.spelling,
-                                        ) + staff_vertical_offset;
-                                        y_raw <= staff_middle_y
-                                    }
-                                };
-                                // Vertical offset to place endpoint at notehead edge
-                                let y_edge = if above {
-                                    -notehead_half_h
-                                } else {
-                                    notehead_half_h
-                                };
-
-                                let in_range = n.start_tick >= system.tick_range.start_tick
-                                    && n.start_tick < system.tick_range.end_tick;
-
-                                if in_range {
-                                    let start_info = match note_lookup.get(n.note_id.as_str()) {
-                                        Some(info) => *info,
-                                        None => continue,
-                                    };
-                                    let (start_x, start_y, _sp, _st) = start_info;
-
-                                    match note_lookup.get(slur_target_id) {
-                                        Some(&(end_x, end_y, _ep, _et)) => {
-                                            // Same-system slur
-                                            let arc_start_x = start_x + notehead_half_w * 0.3;
-                                            let arc_end_x = end_x - notehead_half_w * 0.3;
-                                            let adj_start_y = start_y + y_edge;
-                                            let adj_end_y = end_y + y_edge;
-                                            let span_x = (arc_end_x - arc_start_x).abs().max(1.0);
-                                            // sqrt scaling: short slurs get proportional curves,
-                                            // long slurs keep visible curvature (~1-2 staff spaces)
-                                            let arc_height =
-                                                (3.5 * span_x.sqrt()).clamp(12.0, 50.0);
-                                            let y_offset =
-                                                if above { -arc_height } else { arc_height };
-                                            let mid_y = (adj_start_y + adj_end_y) / 2.0 + y_offset;
-
-                                            slur_arcs.push(types::TieArc {
-                                                start: types::Point {
-                                                    x: arc_start_x,
-                                                    y: adj_start_y,
-                                                },
-                                                end: types::Point {
-                                                    x: arc_end_x,
-                                                    y: adj_end_y,
-                                                },
-                                                cp1: types::Point {
-                                                    x: arc_start_x + span_x * 0.2,
-                                                    y: mid_y,
-                                                },
-                                                cp2: types::Point {
-                                                    x: arc_end_x - span_x * 0.2,
-                                                    y: mid_y,
-                                                },
-                                                above,
-                                                note_id_start: n.note_id.clone(),
-                                                note_id_end: slur_target_id.to_string(),
-                                            });
-                                        }
-                                        None => {
-                                            // Cross-system outgoing slur
-                                            let arc_start_x = start_x + notehead_half_w * 0.3;
-                                            let arc_end_x = system_right_edge;
-                                            let adj_start_y = start_y + y_edge;
-                                            let span_x = arc_end_x - arc_start_x;
-                                            let arc_height =
-                                                (3.5 * span_x.sqrt()).clamp(12.0, 50.0);
-                                            let y_offset =
-                                                if above { -arc_height } else { arc_height };
-                                            let mid_y = adj_start_y + y_offset;
-
-                                            slur_arcs.push(types::TieArc {
-                                                start: types::Point {
-                                                    x: arc_start_x,
-                                                    y: adj_start_y,
-                                                },
-                                                end: types::Point {
-                                                    x: arc_end_x,
-                                                    y: adj_start_y,
-                                                },
-                                                cp1: types::Point {
-                                                    x: arc_start_x + span_x * 0.2,
-                                                    y: mid_y,
-                                                },
-                                                cp2: types::Point {
-                                                    x: arc_end_x - span_x * 0.2,
-                                                    y: mid_y,
-                                                },
-                                                above,
-                                                note_id_start: n.note_id.clone(),
-                                                note_id_end: slur_target_id.to_string(),
-                                            });
-                                        }
-                                    }
-                                } else if n.start_tick < system.tick_range.start_tick
-                                    && system_note_ids.contains(slur_target_id)
-                                {
-                                    // Cross-system incoming slur
-                                    let (end_x, end_y, _ep, _et) =
-                                        *note_lookup.get(slur_target_id).unwrap();
-                                    let adj_end_y = end_y + y_edge;
-                                    let arc_end_x = end_x - notehead_half_w * 0.3;
-                                    let incoming_offset = 3.0 * notehead_half_w;
-                                    let arc_start_x = (unified_left_margin - incoming_offset)
-                                        .min(arc_end_x - 20.0)
-                                        .max(0.0);
-                                    let span_x = (arc_end_x - arc_start_x).max(1.0);
-                                    let arc_height = (3.5 * span_x.sqrt()).clamp(12.0, 50.0);
-                                    let y_offset = if above { -arc_height } else { arc_height };
-                                    let mid_y = adj_end_y + y_offset;
-
-                                    slur_arcs.push(types::TieArc {
-                                        start: types::Point {
-                                            x: arc_start_x,
-                                            y: adj_end_y,
-                                        },
-                                        end: types::Point {
-                                            x: arc_end_x,
-                                            y: adj_end_y,
-                                        },
-                                        cp1: types::Point {
-                                            x: arc_start_x + span_x * 0.2,
-                                            y: mid_y,
-                                        },
-                                        cp2: types::Point {
-                                            x: arc_end_x - span_x * 0.2,
-                                            y: mid_y,
-                                        },
-                                        above,
-                                        note_id_start: n.note_id.clone(),
-                                        note_id_end: slur_target_id.to_string(),
-                                    });
-                                }
-                            }
-                        }
-                    }
-                }
+                // Render annotation elements (ledger lines, dots, ties, slurs)
+                let ann = annotations::render_annotations(
+                    staff_data,
+                    &system.tick_range,
+                    system.index,
+                    system.bounding_box.width,
+                    staff_vertical_offset,
+                    unified_left_margin,
+                    config.units_per_space,
+                    &note_positions,
+                );
 
                 // Create staff with batched glyphs and structural glyphs
                 let staff = Staff {
@@ -1407,10 +518,10 @@ pub fn compute_layout(score: &serde_json::Value, config: &LayoutConfig) -> Globa
                     glyph_runs,
                     structural_glyphs,
                     bar_lines,
-                    ledger_lines,
-                    notation_dots,
-                    tie_arcs,
-                    slur_arcs,
+                    ledger_lines: ann.ledger_lines,
+                    notation_dots: ann.notation_dots,
+                    tie_arcs: ann.tie_arcs,
+                    slur_arcs: ann.slur_arcs,
                 };
 
                 staves.push(staff);
@@ -1424,7 +535,11 @@ pub fn compute_layout(score: &serde_json::Value, config: &LayoutConfig) -> Globa
             };
 
             let bracket_glyph = if instrument.staves.len() > 1 {
-                Some(create_bracket_glyph(&staves, &bracket_type, config))
+                Some(staff_groups::create_bracket_glyph(
+                    &staves,
+                    &bracket_type,
+                    config,
+                ))
             } else {
                 None
             };
@@ -1507,122 +622,17 @@ pub fn compute_layout(score: &serde_json::Value, config: &LayoutConfig) -> Globa
         system.bounding_box.width = content_width;
 
         // Join measure barlines within each multi-staff group.
-        // Standard engraving: internal barlines span from the top line of the
-        // first staff to the bottom line of the last staff in the group.
-        // We adjust the y_start/y_end of each barline segment on staves[0] to
-        // cover the full group height, collect repeat dots from all staves onto
-        // staves[0], then clear barlines from staves[1..].
-        for staff_group in &mut system.staff_groups {
-            if staff_group.staves.len() >= 2 {
-                let top_y = staff_group.staves[0].staff_lines[0].y_position;
-                let bottom_y = staff_group.staves.last().unwrap().staff_lines[4].y_position;
-
-                // Collect repeat dots from subsequent staves, keyed by barline X
-                // so they can be merged onto the matching barline on staves[0].
-                let mut extra_dots: std::collections::HashMap<i32, Vec<RepeatDotPosition>> =
-                    std::collections::HashMap::new();
-                for staff in staff_group.staves[1..].iter() {
-                    for bar_line in &staff.bar_lines {
-                        if bar_line.dots.is_empty() {
-                            continue;
-                        }
-                        let x_key = bar_line
-                            .segments
-                            .first()
-                            .map(|s| (s.x_position * 10.0) as i32)
-                            .unwrap_or(0);
-                        extra_dots
-                            .entry(x_key)
-                            .or_default()
-                            .extend_from_slice(&bar_line.dots);
-                    }
-                }
-
-                // Extend staves[0] barlines to span the full group height.
-                for bar_line in &mut staff_group.staves[0].bar_lines {
-                    for segment in &mut bar_line.segments {
-                        segment.y_start = top_y;
-                        segment.y_end = bottom_y;
-                    }
-                    let x_key = bar_line
-                        .segments
-                        .first()
-                        .map(|s| (s.x_position * 10.0) as i32)
-                        .unwrap_or(0);
-                    if let Some(more_dots) = extra_dots.remove(&x_key) {
-                        bar_line.dots.extend(more_dots);
-                    }
-                }
-
-                // Remove barlines from subsequent staves (merged into staves[0]).
-                for staff in staff_group.staves[1..].iter_mut() {
-                    staff.bar_lines.clear();
-                }
-            }
-        }
+        staff_groups::join_multi_staff_barlines(&mut system.staff_groups);
 
         // Add a system-end barline at the justified right edge for every
-        // staff group.  For multi-staff groups the barline spans the full
-        // group height; for single-staff groups it spans that staff.
-        // The barline type reflects the last measure in this system (Final
-        // for the score's last measure, repeat types if applicable, Single
-        // otherwise).
-        let end_bar_type = {
-            let last_measure = measure_infos
-                .iter()
-                .rfind(|m| m.end_tick == system.tick_range.end_tick);
-            match last_measure {
-                Some(m) => {
-                    let next_start_repeat = measure_infos
-                        .iter()
-                        .find(|mi| mi.start_tick == m.end_tick)
-                        .map(|mi| mi.start_repeat)
-                        .unwrap_or(false);
-                    if m.end_repeat && next_start_repeat {
-                        BarLineType::RepeatBoth
-                    } else if m.end_repeat {
-                        BarLineType::RepeatEnd
-                    } else if next_start_repeat {
-                        BarLineType::RepeatStart
-                    } else {
-                        let is_last =
-                            measure_infos.last().map(|mi| mi.end_tick) == Some(m.end_tick);
-                        if is_last {
-                            BarLineType::Final
-                        } else {
-                            BarLineType::Single
-                        }
-                    }
-                }
-                None => BarLineType::Single,
-            }
-        };
-        for staff_group in &mut system.staff_groups {
-            let top_y = staff_group.staves.first().unwrap().staff_lines[0].y_position;
-            let bottom_y = staff_group.staves.last().unwrap().staff_lines[4].y_position;
-            let segments = create_bar_line_segments(content_width, top_y, bottom_y, &end_bar_type);
-            let dots =
-                compute_repeat_dots(content_width, top_y, config.units_per_space, &end_bar_type);
-            // For multi-staff groups, also add repeat dots for lower staves.
-            let mut all_dots = dots;
-            if staff_group.staves.len() >= 2 {
-                for staff in &staff_group.staves[1..] {
-                    let staff_top = staff.staff_lines[0].y_position;
-                    all_dots.extend(compute_repeat_dots(
-                        content_width,
-                        staff_top,
-                        config.units_per_space,
-                        &end_bar_type,
-                    ));
-                }
-            }
-            let end_barline = BarLine {
-                segments,
-                bar_type: end_bar_type,
-                dots: all_dots,
-            };
-            staff_group.staves[0].bar_lines.push(end_barline);
-        }
+        // staff group.
+        barlines::render_system_barlines(
+            &mut system.staff_groups,
+            &measure_infos,
+            system.tick_range.end_tick,
+            content_width,
+            config.units_per_space,
+        );
 
         // Update system height to include collision-avoidance extra spacing
         system.bounding_box.height += total_collision_extra;
@@ -1755,1841 +765,9 @@ pub fn compute_layout(score: &serde_json::Value, config: &LayoutConfig) -> Globa
     }
 }
 
-/// Extract measures from CompiledScore JSON
-///
-/// Returns a Vec where each element is a Vec of note durations for that measure.
-/// For multi-staff instruments, notes at the same timing position are counted only once
-/// (e.g., treble + bass notes sounding together take one horizontal space).
-///
-/// Returns a Vec where each element is `(note_durations, rest_durations)` for that measure.
-fn extract_measures(
-    score: &serde_json::Value,
-    ticks_per_measure: u32,
-    pickup_ticks: u32,
-    measure_end_ticks: &[u32],
-) -> Vec<(Vec<u32>, Vec<u32>)> {
-    let mut note_measures: Vec<Vec<u32>> = Vec::new();
-    let mut rest_measures: Vec<Vec<u32>> = Vec::new();
-
-    // Extract notes from all instruments
-    if let Some(instruments) = score["instruments"].as_array() {
-        for instrument in instruments {
-            if let Some(staves) = instrument["staves"].as_array() {
-                // Collect all unique timing positions across all staves
-                // (treble + bass notes at same tick = one horizontal position)
-                // Map: tick -> max duration at that tick (use max so wider notes win)
-                let mut all_notes_by_measure: std::collections::HashMap<
-                    usize,
-                    std::collections::HashMap<u32, u32>,
-                > = std::collections::HashMap::new();
-                let mut all_rests_by_measure: std::collections::HashMap<
-                    usize,
-                    std::collections::HashMap<u32, u32>,
-                > = std::collections::HashMap::new();
-
-                for staff in staves {
-                    if let Some(voices) = staff["voices"].as_array() {
-                        for voice in voices {
-                            // --- Notes ---
-                            // Try both "interval_events" (Score format) and "notes" (converted format)
-                            let notes_array = voice["interval_events"]
-                                .as_array()
-                                .or_else(|| voice["notes"].as_array());
-
-                            if let Some(notes) = notes_array {
-                                for note in notes {
-                                    // Support multiple field name formats:
-                                    // Format 1 (Score): start_tick, duration_ticks
-                                    // Format 2 (LayoutView): tick, duration
-                                    // Format 3 (nested): start_tick.value
-                                    let start_tick = note["start_tick"]
-                                        .as_u64()
-                                        .or_else(|| note["tick"].as_u64())
-                                        .or_else(|| note["start_tick"]["value"].as_u64())
-                                        .unwrap_or(0)
-                                        as u32;
-
-                                    let duration = note["duration_ticks"]
-                                        .as_u64()
-                                        .or_else(|| note["duration"].as_u64())
-                                        .unwrap_or(960)
-                                        as u32;
-
-                                    // Determine which measure this note belongs to
-                                    let measure_index = actual_tick_to_measure(
-                                        start_tick,
-                                        measure_end_ticks,
-                                        pickup_ticks,
-                                        ticks_per_measure,
-                                    );
-
-                                    // Track tick → duration (keep max duration at each tick position)
-                                    let entry = all_notes_by_measure
-                                        .entry(measure_index)
-                                        .or_default()
-                                        .entry(start_tick)
-                                        .or_insert(0);
-                                    *entry = (*entry).max(duration);
-                                }
-                            }
-
-                            // --- Rests ---
-                            if let Some(rest_events) = voice["rest_events"].as_array() {
-                                for rest in rest_events {
-                                    let start_tick = rest["start_tick"]
-                                        .as_u64()
-                                        .or_else(|| rest["start_tick"]["value"].as_u64())
-                                        .unwrap_or(0)
-                                        as u32;
-
-                                    let duration =
-                                        rest["duration_ticks"].as_u64().unwrap_or(960) as u32;
-
-                                    let measure_index = actual_tick_to_measure(
-                                        start_tick,
-                                        measure_end_ticks,
-                                        pickup_ticks,
-                                        ticks_per_measure,
-                                    );
-
-                                    let entry = all_rests_by_measure
-                                        .entry(measure_index)
-                                        .or_default()
-                                        .entry(start_tick)
-                                        .or_insert(0);
-                                    *entry = (*entry).max(duration);
-                                }
-                            }
-                        }
-                    }
-                }
-
-                // Convert tick→duration maps to flat duration lists for compute_measure_width
-                let max_measure = all_notes_by_measure
-                    .keys()
-                    .chain(all_rests_by_measure.keys())
-                    .copied()
-                    .max()
-                    .unwrap_or(0);
-
-                for measure_index in 0..=max_measure {
-                    while note_measures.len() <= measure_index {
-                        note_measures.push(Vec::new());
-                    }
-                    while rest_measures.len() <= measure_index {
-                        rest_measures.push(Vec::new());
-                    }
-
-                    if let Some(tick_durations) = all_notes_by_measure.get(&measure_index) {
-                        for dur in tick_durations.values() {
-                            note_measures[measure_index].push(*dur);
-                        }
-                    }
-                    if let Some(tick_durations) = all_rests_by_measure.get(&measure_index) {
-                        for dur in tick_durations.values() {
-                            rest_measures[measure_index].push(*dur);
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    // If no measures found, return empty default measures
-    if note_measures.is_empty() && rest_measures.is_empty() {
-        // 10 measures with 4 quarter notes each (no rests)
-        (0..10).map(|_| (vec![960; 4], Vec::new())).collect()
-    } else {
-        let len = note_measures.len().max(rest_measures.len());
-        note_measures.resize(len, Vec::new());
-        rest_measures.resize(len, Vec::new());
-        note_measures.into_iter().zip(rest_measures).collect()
-    }
-}
-
-/// Represents an instrument with its staves extracted from CompiledScore
-#[derive(Debug, Clone)]
-struct InstrumentData {
-    id: String,
-    name: String,
-    staves: Vec<StaffData>,
-}
-
-/// Represents a staff with voices and notes
-#[derive(Debug, Clone)]
-struct StaffData {
-    voices: Vec<VoiceData>,
-    clef: String,         // e.g., "Treble", "Bass", "Alto", "Tenor"
-    time_numerator: u8,   // e.g., 4 for 4/4 time
-    time_denominator: u8, // e.g., 4 for 4/4 time
-    key_sharps: i8,       // Initial key: positive for sharps, negative for flats, 0 for C major
-    /// Key signature changes sorted by tick. Empty if no mid-piece changes.
-    key_signature_events: Vec<(u32, i8)>,
-    /// Clef changes sorted by tick. Empty if no mid-piece changes.
-    clef_events: Vec<(u32, String)>,
-}
-
-impl StaffData {
-    /// Get the active key signature (sharps count) at a given tick.
-    fn get_key_at_tick(&self, tick: u32) -> i8 {
-        if self.key_signature_events.is_empty() {
-            return self.key_sharps;
-        }
-        // Find the last event whose tick <= the query tick
-        let mut result = self.key_sharps;
-        for &(event_tick, sharps) in &self.key_signature_events {
-            if event_tick <= tick {
-                result = sharps;
-            } else {
-                break;
-            }
-        }
-        result
-    }
-
-    /// Get the active clef at a given tick.
-    fn get_clef_at_tick(&self, tick: u32) -> &str {
-        if self.clef_events.is_empty() {
-            return &self.clef;
-        }
-        let mut result = &self.clef;
-        for (event_tick, clef) in &self.clef_events {
-            if *event_tick <= tick {
-                result = clef;
-            } else {
-                break;
-            }
-        }
-        result
-    }
-}
-
-/// Represents a voice with interval events (notes) and rest events
-#[derive(Debug, Clone)]
-struct VoiceData {
-    notes: Vec<NoteEvent>,
-    rests: Vec<RestLayoutEvent>,
-}
-
-/// Rest event extracted from ScoreDto JSON for layout processing
-#[derive(Debug, Clone)]
-pub(super) struct RestLayoutEvent {
-    start_tick: u32,
-    duration_ticks: u32,
-    note_type: Option<String>,
-    /// MusicXML voice number (1-indexed): odd = Voice 1 (up), even = Voice 2 (down)
-    voice: usize,
-}
-
-/// Note data tuple: (pitch, start_tick, duration_ticks, spelling, staccato, dot_count)
-///
-/// Spelling is an optional (step_letter, alter) pair from MusicXML,
-/// e.g. ('E', -1) for Eb, ('D', 1) for D#.
-pub type NoteData = (u8, u32, u32, Option<(char, i8)>, bool, u8);
-
-/// Represents a single note event
-#[derive(Debug, Clone)]
-struct NoteEvent {
-    pitch: u8,
-    start_tick: u32,
-    duration_ticks: u32,
-    /// Explicit spelling from MusicXML: (step_letter, alter) e.g. ('E', -1) for Eb
-    spelling: Option<(char, i8)>,
-    /// Beam annotations from MusicXML import (empty = needs algorithmic grouping)
-    beam_info: Vec<(u8, String)>, // (beam_level, beam_type_string)
-    /// Staccato articulation
-    staccato: bool,
-    /// Number of augmentation dots (0, 1, or 2)
-    dot_count: u8,
-    /// Domain Note ID (UUID string) for tie arc linking
-    note_id: String,
-    /// If this note starts/continues a tie, the ID of the next tied note
-    tie_next: Option<String>,
-    /// If a slur starts on this note, the ID of the note where it ends
-    slur_next: Option<String>,
-    /// Slur direction from MusicXML: Some(true)=above, Some(false)=below, None=auto
-    slur_above: Option<bool>,
-}
-
-/// Extract instruments from CompiledScore JSON
-fn extract_instruments(
-    score: &serde_json::Value,
-    global_time_numerator: u32,
-    global_time_denominator: u32,
-) -> Vec<InstrumentData> {
-    let mut instruments = Vec::new();
-
-    if let Some(instruments_array) = score["instruments"].as_array() {
-        for instrument in instruments_array {
-            let id = instrument["id"].as_str().unwrap_or("unknown").to_string();
-            let name = instrument["name"]
-                .as_str()
-                .unwrap_or("Instrument")
-                .to_string();
-            let mut staves = Vec::new();
-
-            if let Some(staves_array) = instrument["staves"].as_array() {
-                for staff in staves_array {
-                    let mut voices = Vec::new();
-
-                    // Extract structural metadata (with defaults)
-                    let clef = staff["clef"].as_str().unwrap_or("Treble").to_string();
-                    // Use time signature from global_structural_events;
-                    // fall back to staff-level for test fixtures that set it there
-                    let time_numerator = if global_time_numerator != 4
-                        || global_time_denominator != 4
-                    {
-                        global_time_numerator as u8
-                    } else {
-                        staff["time_signature"]["numerator"]
-                            .as_u64()
-                            .unwrap_or(global_time_numerator as u64) as u8
-                    };
-                    let time_denominator = if global_time_numerator != 4
-                        || global_time_denominator != 4
-                    {
-                        global_time_denominator as u8
-                    } else {
-                        staff["time_signature"]["denominator"]
-                            .as_u64()
-                            .unwrap_or(global_time_denominator as u64) as u8
-                    };
-                    let key_sharps = staff["key_signature"]["sharps"].as_i64().unwrap_or(0) as i8;
-
-                    if let Some(voices_array) = staff["voices"].as_array() {
-                        for voice in voices_array {
-                            let mut notes = Vec::new();
-
-                            // T008-T009: Support both "notes" (LayoutView format) and "interval_events" (CompiledScore format)
-                            // Check "notes" first for frontend fixtures, fall back to "interval_events" for backward compatibility
-
-                            let note_array = voice["notes"]
-                                .as_array()
-                                .or_else(|| voice["interval_events"].as_array());
-
-                            if let Some(notes_data) = note_array {
-                                for note_item in notes_data {
-                                    // Handle both formats:
-                                    // Format 1 (notes): {tick: 0, duration: 960, pitch: 60}
-                                    // Format 2 (interval_events): {start_tick: {value: 0}, duration_ticks: 960, pitch: {value: 60}}
-
-                                    let pitch = if let Some(p) = note_item["pitch"].as_u64() {
-                                        p as u8 // Format 1: direct value
-                                    } else {
-                                        note_item["pitch"]["value"].as_u64().unwrap_or(60) as u8 // Format 2: nested
-                                    };
-
-                                    let start_tick = if let Some(t) = note_item["tick"].as_u64() {
-                                        t as u32 // Format 1: "tick"
-                                    } else {
-                                        note_item["start_tick"]
-                                            .as_u64()
-                                            .or_else(|| note_item["start_tick"]["value"].as_u64())
-                                            .unwrap_or(0)
-                                            as u32 // Format 2: plain int or nested
-                                    };
-
-                                    let duration_ticks = if let Some(d) =
-                                        note_item["duration"].as_u64()
-                                    {
-                                        d as u32 // Format 1: "duration"
-                                    } else {
-                                        note_item["duration_ticks"].as_u64().unwrap_or(960) as u32 // Format 2
-                                    };
-
-                                    // Extract optional note spelling (step + alter) from MusicXML
-                                    let spelling = note_item["spelling"]["step"]
-                                        .as_str()
-                                        .and_then(|s| s.chars().next())
-                                        .and_then(|step| {
-                                            note_item["spelling"]["alter"]
-                                                .as_i64()
-                                                .map(|alter| (step, alter as i8))
-                                        });
-
-                                    notes.push(NoteEvent {
-                                        pitch,
-                                        start_tick,
-                                        duration_ticks,
-                                        spelling,
-                                        beam_info: {
-                                            let mut beams = Vec::new();
-                                            if let Some(beam_array) = note_item["beams"].as_array()
-                                            {
-                                                for beam_item in beam_array {
-                                                    let number =
-                                                        beam_item["number"].as_u64().unwrap_or(1)
-                                                            as u8;
-                                                    let beam_type = beam_item["beam_type"]
-                                                        .as_str()
-                                                        .unwrap_or("")
-                                                        .to_string();
-                                                    beams.push((number, beam_type));
-                                                }
-                                            }
-                                            beams
-                                        },
-                                        staccato: note_item["staccato"].as_bool().unwrap_or(false),
-                                        dot_count: note_item["dot_count"].as_u64().unwrap_or(0)
-                                            as u8,
-                                        note_id: note_item["id"].as_str().unwrap_or("").to_string(),
-                                        tie_next: note_item["tie_next"]
-                                            .as_str()
-                                            .map(|s| s.to_string()),
-                                        slur_next: note_item["slur_next"]
-                                            .as_str()
-                                            .map(|s| s.to_string()),
-                                        slur_above: note_item["slur_above"].as_bool(),
-                                    });
-                                }
-                            }
-
-                            voices.push(VoiceData {
-                                notes,
-                                rests: {
-                                    let mut rests = Vec::new();
-                                    if let Some(rest_events) = voice["rest_events"].as_array() {
-                                        for rest_item in rest_events {
-                                            let start_tick = rest_item["start_tick"]
-                                                .as_u64()
-                                                .or_else(|| {
-                                                    rest_item["start_tick"]["value"].as_u64()
-                                                })
-                                                .unwrap_or(0)
-                                                as u32;
-                                            let duration_ticks =
-                                                rest_item["duration_ticks"].as_u64().unwrap_or(960)
-                                                    as u32;
-                                            let note_type = rest_item["note_type"]
-                                                .as_str()
-                                                .map(|s| s.to_string());
-                                            let voice_num =
-                                                rest_item["voice"].as_u64().unwrap_or(1) as usize;
-                                            rests.push(RestLayoutEvent {
-                                                start_tick,
-                                                duration_ticks,
-                                                note_type,
-                                                voice: voice_num,
-                                            });
-                                        }
-                                    }
-                                    rests
-                                },
-                            });
-                        }
-                    }
-
-                    // Parse key signature change events (for mid-piece key changes)
-                    let mut key_signature_events: Vec<(u32, i8)> = Vec::new();
-                    if let Some(events_array) = staff["key_signature_events"].as_array() {
-                        for ev in events_array {
-                            let tick = ev["tick"].as_u64().unwrap_or(0) as u32;
-                            let sharps = ev["sharps"].as_i64().unwrap_or(0) as i8;
-                            key_signature_events.push((tick, sharps));
-                        }
-                    }
-                    // Also extract from staff_structural_events (DTO path)
-                    if let Some(events_array) = staff["staff_structural_events"].as_array() {
-                        for ev in events_array {
-                            if let Some(ks_obj) = ev.get("KeySignature") {
-                                let tick = ks_obj["tick"]
-                                    .as_u64()
-                                    .or_else(|| ks_obj["tick"]["value"].as_u64())
-                                    .unwrap_or(0) as u32;
-                                let sharps = ks_obj["sharps"].as_i64().unwrap_or(0) as i8;
-                                key_signature_events.push((tick, sharps));
-                            }
-                        }
-                    }
-                    key_signature_events.sort_by_key(|&(t, _)| t);
-                    key_signature_events.dedup_by_key(|&mut (t, _)| t);
-
-                    // Parse clef change events (for mid-piece clef changes)
-                    // Clef events may be stored in two places:
-                    //  1. A flat "clef_events" array (legacy path)
-                    //  2. Inside "staff_structural_events" as {"Clef": {"tick":..,"clef":..}}
-                    let mut clef_events: Vec<(u32, String)> = Vec::new();
-                    if let Some(events_array) = staff["clef_events"].as_array() {
-                        for ev in events_array {
-                            let tick = ev["tick"]
-                                .as_u64()
-                                .or_else(|| ev["tick"]["value"].as_u64())
-                                .unwrap_or(0) as u32;
-                            let clef_name = ev["clef"].as_str().unwrap_or("Treble").to_string();
-                            clef_events.push((tick, clef_name));
-                        }
-                    }
-                    // Also extract from staff_structural_events (DTO path)
-                    if let Some(events_array) = staff["staff_structural_events"].as_array() {
-                        for ev in events_array {
-                            if let Some(clef_obj) = ev.get("Clef") {
-                                let tick = clef_obj["tick"]
-                                    .as_u64()
-                                    .or_else(|| clef_obj["tick"]["value"].as_u64())
-                                    .unwrap_or(0) as u32;
-                                let clef_name =
-                                    clef_obj["clef"].as_str().unwrap_or("Treble").to_string();
-                                clef_events.push((tick, clef_name));
-                            }
-                        }
-                    }
-                    clef_events.sort_by_key(|(t, _)| *t);
-                    clef_events.dedup_by_key(|(t, _)| *t);
-
-                    staves.push(StaffData {
-                        voices,
-                        clef,
-                        time_numerator,
-                        time_denominator,
-                        key_sharps,
-                        key_signature_events,
-                        clef_events,
-                    });
-                }
-            }
-
-            instruments.push(InstrumentData { id, name, staves });
-        }
-    }
-
-    instruments
-}
-
-/// Compute unified note positions across all staves in a system
-///
-/// For multi-instrument scores, notes at the same tick must align
-/// horizontally across ALL staves and instruments. This function collects
-/// all unique tick positions from every staff and computes a unified spacing map.
-///
-/// # Arguments
-/// * `staves` - All staves across all instruments in this system
-/// * `tick_range` - Tick range for this system
-/// * `system_width` - Total system width in logical units
-/// * `left_margin` - Left margin for note area
-/// * `spacing_config` - Spacing configuration
-///
-/// # Returns
-/// HashMap mapping tick positions to x-coordinates (Principle VI: Layout Engine Authority)
-fn compute_unified_note_positions(
-    staves: &[&StaffData],
-    tick_range: &TickRange,
-    system_width: f32,
-    left_margin: f32,
-    spacing_config: &spacer::SpacingConfig,
-    ticks_per_measure: u32,
-    clef_change_ticks: &std::collections::HashSet<u32>,
-) -> HashMap<u32, f32> {
-    // Collect all unique (tick, duration) pairs from all staves — notes AND
-    // sub-beat rests. Full-measure rests are excluded because they are centred
-    // within their measure via measure_x_bounds, not time-proportionally spaced.
-    let mut tick_durations: Vec<(u32, u32)> = Vec::new();
-
-    for staff_data in staves {
-        for voice in &staff_data.voices {
-            for note in &voice.notes {
-                if note.start_tick >= tick_range.start_tick && note.start_tick < tick_range.end_tick
-                {
-                    tick_durations.push((note.start_tick, note.duration_ticks));
-                }
-            }
-            for rest in &voice.rests {
-                if rest.start_tick >= tick_range.start_tick
-                    && rest.start_tick < tick_range.end_tick
-                    && rest.duration_ticks < ticks_per_measure
-                {
-                    tick_durations.push((rest.start_tick, rest.duration_ticks));
-                }
-            }
-        }
-    }
-
-    if tick_durations.is_empty() {
-        return HashMap::new();
-    }
-
-    // Sort by tick position and remove duplicates
-    tick_durations.sort_by_key(|(tick, _)| *tick);
-    tick_durations.dedup_by_key(|(tick, _)| *tick);
-
-    // Calculate cumulative spacing based on durations
-    let mut cumulative_spacing = Vec::new();
-    let mut current_position = 0.0;
-    let mut last_tick = tick_range.start_tick;
-
-    for (start_tick, duration_ticks) in &tick_durations {
-        if *start_tick > last_tick {
-            let gap_duration = (*start_tick - last_tick).min(*duration_ticks);
-            let mut gap = spacer::compute_note_spacing(gap_duration, spacing_config);
-            // Add extra visual space at measure boundaries so barlines have
-            // breathing room on both sides. Skip the very first note in the
-            // system (no preceding barline) but boost every other measure start.
-            if *start_tick > tick_range.start_tick
-                && ticks_per_measure > 0
-                && *start_tick % ticks_per_measure == 0
-            {
-                gap += 25.0;
-            }
-            // Add extra space for clef changes so the clef glyph
-            // doesn't collide with the notes that follow it.
-            if clef_change_ticks.contains(start_tick) {
-                gap += 50.0;
-            }
-            current_position += gap;
-        }
-        cumulative_spacing.push(current_position);
-        last_tick = *start_tick;
-    }
-
-    // Calculate total natural width
-    // Add clearance space after last note for notehead width (≈20 units = half the notehead)
-    // plus barline spacing (≈10 units) = 30 units total
-    let end_clearance = 30.0;
-    let total_natural_width = if let Some(&last_pos) = cumulative_spacing.last() {
-        let (_, last_duration) = tick_durations.last().unwrap();
-        last_pos + spacer::compute_note_spacing(*last_duration, spacing_config) + end_clearance
-    } else {
-        spacing_config.minimum_spacing + end_clearance
-    };
-
-    // Scale positions to fill available width (system justification).
-    // This ensures all systems have equal width and measures are evenly distributed.
-    let available_width = system_width - left_margin;
-    let scale_factor = if total_natural_width > 0.0 {
-        available_width / total_natural_width
-    } else {
-        1.0
-    };
-
-    // Build position map: tick -> x_position
-    let mut position_map = HashMap::new();
-    for (i, (tick, _)) in tick_durations.iter().enumerate() {
-        let x_position = left_margin + (cumulative_spacing[i] * scale_factor);
-        position_map.insert(*tick, x_position);
-    }
-
-    position_map
-}
-
-/// Position glyphs for a single staff within a system's tick range
-///
-/// Uses pre-computed unified note positions to ensure horizontal alignment
-/// across all staves in a staff group (e.g., piano treble/bass).
-///
-/// # Arguments
-/// * `staff_data` - Staff data containing voices and notes
-/// * `tick_range` - Tick range for this system
-/// * `units_per_space` - Scaling factor (20 units = 1 staff space)
-/// * `instrument_id` - Instrument identifier
-/// * `staff_index` - Index of this staff within the instrument
-/// * `staff_vertical_offset` - Vertical offset for this staff
-/// * `note_positions` - Pre-computed tick -> x_position map (unified across staves)
-/// * `left_margin` - Left margin (clef + key sig + time sig width) for rest fallback
-#[allow(clippy::too_many_arguments)]
-fn position_glyphs_for_staff(
-    staff_data: &StaffData,
-    tick_range: &TickRange,
-    units_per_space: f32,
-    instrument_id: &str,
-    staff_index: usize,
-    staff_vertical_offset: f32,
-    note_positions: &HashMap<u32, f32>,
-    left_margin: f32,
-    ticks_per_measure: u32,
-    measure_x_bounds: &HashMap<u32, (f32, f32)>,
-    pickup_ticks: u32,
-) -> Vec<Glyph> {
-    let mut all_glyphs = Vec::new();
-    let num_voices = staff_data.voices.len();
-
-    for (voice_index, voice) in staff_data.voices.iter().enumerate() {
-        // Multi-voice stem rule: voice 0 always stems up, voice 1+ always stems down
-        let forced_stem_down: Option<bool> = if num_voices > 1 {
-            Some(voice_index > 0)
-        } else {
-            None
-        };
-
-        // Filter notes that fall within this system's tick range
-        let notes_in_range: Vec<NoteData> = voice
-            .notes
-            .iter()
-            .filter(|note| {
-                note.start_tick >= tick_range.start_tick && note.start_tick < tick_range.end_tick
-            })
-            .map(|note| {
-                (
-                    note.pitch,
-                    note.start_tick,
-                    note.duration_ticks,
-                    note.spelling,
-                    note.staccato,
-                    note.dot_count,
-                )
-            })
-            .collect();
-
-        if notes_in_range.is_empty() {
-            continue;
-        }
-
-        // Use pre-computed positions from unified spacing (Principle VI)
-        let horizontal_offsets: Vec<f32> = notes_in_range
-            .iter()
-            .map(|(_, start_tick, _, _, _, _)| *note_positions.get(start_tick).unwrap_or(&0.0))
-            .collect();
-
-        // Position noteheads using positioner module
-        // First, compute beam groups to determine which notes are beamed
-        let voice_notes_in_range: Vec<&NoteEvent> = voice
-            .notes
-            .iter()
-            .filter(|note| {
-                note.start_tick >= tick_range.start_tick && note.start_tick < tick_range.end_tick
-            })
-            .collect();
-
-        // Compute per-note clef types based on active clef at each note's tick
-        let note_clefs: Vec<&str> = notes_in_range
-            .iter()
-            .map(|(_, start_tick, _, _, _, _)| staff_data.get_clef_at_tick(*start_tick))
-            .collect();
-
-        // Build beamable notes with beam info for beam group analysis
-        let beamable_for_analysis_raw: Vec<beams::BeamableNote> = voice_notes_in_range
-            .iter()
-            .enumerate()
-            .filter(|(_, note)| {
-                // Include eighth notes and shorter, plus any note with explicit
-                // MusicXML beam annotations (e.g. dotted eighths in beam groups)
-                note.duration_ticks <= 480 || !note.beam_info.is_empty()
-            })
-            .map(|(i, note)| {
-                let notehead_x = horizontal_offsets[i];
-                let notehead_y = positioner::pitch_to_y(note.pitch, note_clefs[i], units_per_space)
-                    + staff_vertical_offset;
-
-                // Convert beam_info from (number, type_string) to beam_types list
-                let beam_types: Vec<String> =
-                    note.beam_info.iter().map(|(_, bt)| bt.clone()).collect();
-                let beam_levels = note.beam_info.len() as u8;
-
-                beams::BeamableNote {
-                    x: notehead_x,
-                    y: notehead_y,
-                    stem_end_y: 0.0, // Will be computed after stems
-                    tick: note.start_tick,
-                    duration_ticks: note.duration_ticks,
-                    beam_levels,
-                    beam_types,
-                    event_index: i, // Local index within this system's voice
-                }
-            })
-            .collect();
-
-        // Deduplicate chord notes: keep only one entry per tick for beam grouping.
-        // Chord notes share a stem, so beaming only needs one note per time position.
-        // For the Y position, use the lowest note (highest Y) for stem-down groups
-        // and the highest note (lowest Y) for stem-up groups — we'll pick the first
-        // occurrence which is sufficient for beam grouping since stem direction is
-        // determined later by the group as a whole.
-        let mut beamable_for_analysis: Vec<beams::BeamableNote> = Vec::new();
-        let mut seen_ticks: std::collections::HashSet<u32> = std::collections::HashSet::new();
-        for note in beamable_for_analysis_raw {
-            if seen_ticks.insert(note.tick) {
-                beamable_for_analysis.push(note);
-            }
-        }
-
-        // Determine beam groups from MusicXML beam data
-        let has_beam_info = beamable_for_analysis
-            .iter()
-            .any(|n| !n.beam_types.is_empty());
-        let beam_groups = if has_beam_info {
-            beams::build_beam_groups_from_musicxml(&beamable_for_analysis)
-        } else {
-            // T041: Algorithmic fallback using time signature-aware beat grouping
-            let groups = beams::group_beamable_by_time_signature(
-                &beamable_for_analysis,
-                staff_data.time_numerator,
-                staff_data.time_denominator,
-            );
-            groups
-                .into_iter()
-                .map(|notes| beams::BeamGroup {
-                    beam_count: 1,
-                    notes,
-                })
-                .collect()
-        };
-
-        // Build set of beamed note indices (indices into notes_in_range)
-        let mut beamed_note_indices = std::collections::HashSet::<usize>::new();
-        // Map tick → ALL note indices at that tick (chords have multiple notes per tick)
-        let mut tick_to_indices: std::collections::HashMap<u32, Vec<usize>> =
-            std::collections::HashMap::new();
-        for (idx, note_data) in notes_in_range.iter().enumerate() {
-            let (_, start_tick, duration_ticks, _, _, _) = note_data;
-            // Include notes ≤ eighth duration or with explicit beam annotations
-            let has_beam = voice_notes_in_range
-                .iter()
-                .any(|n| n.start_tick == *start_tick && !n.beam_info.is_empty());
-            if *duration_ticks <= 480 || has_beam {
-                tick_to_indices.entry(*start_tick).or_default().push(idx);
-            }
-        }
-        for group in &beam_groups {
-            // T050: Skip degenerate single-note groups — they render with flags, not beams
-            if group.notes.len() < 2 {
-                continue;
-            }
-            for note in &group.notes {
-                if let Some(indices) = tick_to_indices.get(&note.tick) {
-                    for &idx in indices {
-                        beamed_note_indices.insert(idx);
-                    }
-                }
-            }
-        }
-
-        // === CHORD DISPLACEMENT ===
-        // Detect chords (multiple notes at the same tick) and:
-        //   1. Compute per-note x offsets for adjacent pairs (one notehead shifts right
-        //      by one full notehead width so the two heads are no longer overlapping).
-        //      Pattern follows standard engraving (Gould "Behind Bars"): from the bottom
-        //      note of a chord upward, alternate left/right positions — the 2nd note in
-        //      each cluster of seconds displaces, the 3rd stays, the 4th displaces, etc.
-        //   2. Mark secondary (non-source) chord members as bare noteheads so only one
-        //      combined note+stem glyph appears per chord.
-
-        // Pre-compute Y positions in the same coordinate space used for adjacency checks.
-        let chord_note_y_positions: Vec<f32> = notes_in_range
-            .iter()
-            .enumerate()
-            .map(|(i, (pitch, _, _, spelling, _, _))| {
-                positioner::pitch_to_y_with_spelling(
-                    *pitch,
-                    note_clefs[i],
-                    units_per_space,
-                    *spelling,
-                ) + staff_vertical_offset
-            })
-            .collect();
-
-        let chord_stem_middle_y = staff_vertical_offset + 1.5 * units_per_space;
-        // A diatonic second spans 0.5 * units_per_space in our coordinate system.
-        // Add a small epsilon to handle floating-point rounding.
-        let chord_adjacent_threshold = 0.5 * units_per_space + 0.01;
-        // Build tick → note-index map for ALL notes (not just beamed).
-        let mut chord_tick_to_indices: std::collections::HashMap<u32, Vec<usize>> =
-            std::collections::HashMap::new();
-        for (idx, (_, start_tick, _, _, _, _)) in notes_in_range.iter().enumerate() {
-            chord_tick_to_indices
-                .entry(*start_tick)
-                .or_default()
-                .push(idx);
-        }
-
-        let mut chord_x_offsets: Vec<f32> = vec![0.0; notes_in_range.len()];
-        // Maps note index → render font_size override for chord noteheads.
-        // Bare noteheads (noteheadBlack/noteheadHalf) are scaled up to visually match
-        // the notehead size inside Bravura's combined glyphs (noteQuarterDown, noteHalfDown).
-        // Scale factors derived from Bravura 1000-upem font metrics:
-        //   noteheadBlack advance=295  vs noteQuarterDown xMax=332 → ×(332/295)≈1.125
-        //   noteheadHalf  advance=300  vs noteHalfDown    xMax=345 → ×(345/300)≈1.150
-        let mut chord_scale_map: std::collections::HashMap<usize, f32> =
-            std::collections::HashMap::new();
-        // Chord stems for non-beamed chords: (x, y_top, y_bottom, event_index)
-        let mut chord_stem_data: Vec<(f32, f32, f32, usize)> = Vec::new();
-
-        for indices in chord_tick_to_indices.values() {
-            if indices.len() < 2 {
-                continue;
-            }
-
-            // Sort by y descending: sorted[0] = highest y = lowest pitch (bottom of chord),
-            // sorted.last() = lowest y = highest pitch (top of chord).
-            let mut sorted: Vec<usize> = indices.clone();
-            sorted.sort_by(|&a, &b| {
-                chord_note_y_positions[b]
-                    .partial_cmp(&chord_note_y_positions[a])
-                    .unwrap_or(std::cmp::Ordering::Equal)
-            });
-
-            // Chord stem direction: forced by voice in multi-voice, otherwise
-            // based on whichever note is furthest from the middle line.
-            let chord_stem_down = if let Some(forced) = forced_stem_down {
-                forced
-            } else {
-                let most_extreme = *sorted
-                    .iter()
-                    .max_by(|&&a, &&b| {
-                        let da = (chord_note_y_positions[a] - chord_stem_middle_y).abs();
-                        let db = (chord_note_y_positions[b] - chord_stem_middle_y).abs();
-                        da.partial_cmp(&db).unwrap_or(std::cmp::Ordering::Equal)
-                    })
-                    .unwrap();
-                chord_note_y_positions[most_extreme] <= chord_stem_middle_y
-            };
-
-            // Anchor note for stem attachment (non-displaced):
-            //   Stem-up  → lowest note (highest y) = sorted[0]
-            //   Stem-down → highest note (lowest y) = sorted.last()
-            let anchor_idx = if chord_stem_down {
-                *sorted.last().unwrap()
-            } else {
-                sorted[0]
-            };
-
-            // Check if chord needs an explicit stem pseudo-glyph.
-            // Quarter notes and longer (except whole notes) have stems but are not beamed.
-            let chord_duration = notes_in_range[sorted[0]].2;
-            let needs_explicit_stem = (960..3840).contains(&chord_duration);
-            let any_beamed = sorted.iter().any(|idx| beamed_note_indices.contains(idx));
-
-            // Scale factor to make bare noteheads match combined-glyph notehead widths.
-            // Whole notes (duration >= 3840) use noteheadWhole in both chord and standalone
-            // paths — no combined glyph exists — so no scaling is needed (scale = 1.0).
-            let notehead_scale: f32 = if chord_duration >= 3840 {
-                1.0 // noteheadWhole — same glyph in chord and standalone
-            } else if chord_duration >= 1920 {
-                345.0 / 300.0 // noteheadHalf → matches noteHalfDown notehead width
-            } else {
-                332.0 / 295.0 // noteheadBlack → matches noteQuarterDown notehead width
-            };
-            let chord_font_size = 80.0 * notehead_scale;
-            let scaled_half_width = stems::Stem::NOTEHEAD_WIDTH * notehead_scale;
-            let chord_displacement = scaled_half_width * 2.0;
-
-            if needs_explicit_stem && !any_beamed {
-                // ALL chord members use scaled bare noteheads + one shared explicit stem.
-                for &idx in &sorted {
-                    chord_scale_map.insert(idx, chord_font_size);
-                }
-
-                // Compute stem geometry using scaled notehead half-width.
-                let visual_y_offset = 0.5 * units_per_space;
-                let bottom_y = chord_note_y_positions[sorted[0]] + visual_y_offset;
-                let top_y = chord_note_y_positions[*sorted.last().unwrap()] + visual_y_offset;
-
-                if chord_stem_down {
-                    // Stem on left side of highest (anchor) note, extends below lowest note
-                    let stem_x = horizontal_offsets[anchor_idx] - scaled_half_width;
-                    chord_stem_data.push((
-                        stem_x,
-                        top_y,
-                        bottom_y + stems::Stem::STEM_LENGTH,
-                        anchor_idx,
-                    ));
-                } else {
-                    // Stem on right side of lowest (anchor) note, extends above highest note
-                    let stem_x = horizontal_offsets[anchor_idx] + scaled_half_width;
-                    chord_stem_data.push((
-                        stem_x,
-                        top_y - stems::Stem::STEM_LENGTH,
-                        bottom_y,
-                        anchor_idx,
-                    ));
-                }
-            } else {
-                // For beamed chords or short-duration chords: only non-anchor notes
-                // become secondary (anchor keeps its combined note+stem glyph).
-                for &idx in &sorted {
-                    if idx != anchor_idx && !beamed_note_indices.contains(&idx) {
-                        chord_scale_map.insert(idx, chord_font_size);
-                    }
-                }
-            }
-
-            // Displacement for adjacent pairs — alternating L/R pattern from the bottom:
-            //   sorted[0] stays (L), sorted[1] displaces if adjacent (R),
-            //   sorted[2] stays if adjacent to sorted[1] (L), and so on.
-            // This ensures the upper note of each adjacent pair moves to the right,
-            // matching standard engraving convention.
-            let mut next_should_displace = true;
-            for i in 1..sorted.len() {
-                let y_lower = chord_note_y_positions[sorted[i - 1]]; // lower pitch (higher y)
-                let y_upper = chord_note_y_positions[sorted[i]]; // higher pitch (lower y)
-                let y_diff = y_lower - y_upper; // positive
-
-                if y_diff <= chord_adjacent_threshold {
-                    if next_should_displace {
-                        chord_x_offsets[sorted[i]] += chord_displacement;
-                        next_should_displace = false;
-                    } else {
-                        next_should_displace = true;
-                    }
-                } else {
-                    // Non-adjacent break: reset so the next adjacent pair starts fresh.
-                    next_should_displace = true;
-                }
-            }
-        }
-
-        // Merge chord x-offsets into horizontal offsets for both noteheads and accidentals.
-        let adjusted_horizontal_offsets: Vec<f32> = horizontal_offsets
-            .iter()
-            .zip(chord_x_offsets.iter())
-            .map(|(x, dx)| x + dx)
-            .collect();
-
-        let glyphs = positioner::position_noteheads(
-            &notes_in_range,
-            &adjusted_horizontal_offsets,
-            &note_clefs,
-            units_per_space,
-            instrument_id,
-            staff_index,
-            voice_index,
-            staff_vertical_offset,
-            &beamed_note_indices,
-            &chord_scale_map,
-            forced_stem_down,
-        );
-
-        all_glyphs.extend(glyphs);
-
-        // Generate explicit stem pseudo-glyphs for non-beamed chords.
-        // These replace the built-in stem that combined note glyphs normally provide.
-        for &(stem_x, y_top, y_bottom, event_index) in &chord_stem_data {
-            let stem_height = y_bottom - y_top;
-            let stem_glyph = Glyph {
-                codepoint: '\u{0000}'.to_string(),
-                position: Point {
-                    x: stem_x,
-                    y: y_top,
-                },
-                bounding_box: BoundingBox {
-                    x: stem_x - (stems::Stem::STEM_THICKNESS / 2.0),
-                    y: y_top,
-                    width: stems::Stem::STEM_THICKNESS,
-                    height: stem_height,
-                },
-                source_reference: SourceReference {
-                    instrument_id: instrument_id.to_string(),
-                    staff_index,
-                    voice_index,
-                    event_index,
-                },
-                font_size: None,
-            };
-            all_glyphs.push(stem_glyph);
-        }
-
-        // Position accidentals based on key signature and measure context
-        let accidental_glyphs = positioner::position_note_accidentals(
-            &notes_in_range,
-            &adjusted_horizontal_offsets,
-            &note_clefs,
-            units_per_space,
-            instrument_id,
-            staff_index,
-            voice_index,
-            staff_vertical_offset,
-            staff_data.key_sharps,
-            ticks_per_measure,
-            &staff_data.key_signature_events,
-            pickup_ticks,
-        );
-
-        all_glyphs.extend(accidental_glyphs);
-
-        // Generate stems and beams for beamed notes
-        // Staff middle line (line 2 of 5) is at y = staff_vertical_offset + 2.0 * units_per_space.
-        // However, pitch_to_y values include a -0.5*ups glyph-centering offset, so we adjust
-        // staff_middle_y to match the same coordinate space for correct stem direction comparison.
-        let staff_middle_y = staff_vertical_offset + 1.5 * units_per_space;
-
-        let mut stem_glyphs = Vec::new();
-
-        // For each beam group, compute uniform stem direction and generate stems + beams
-        for group in &beam_groups {
-            if group.notes.len() < 2 {
-                continue; // Skip degenerate groups (single note → use flag)
-            }
-
-            // T033: Compute uniform stem direction for the group.
-            // Multi-voice override takes priority; otherwise use majority rule.
-            let group_direction = if let Some(forced) = forced_stem_down {
-                if forced {
-                    stems::StemDirection::Down
-                } else {
-                    stems::StemDirection::Up
-                }
-            } else {
-                beams::compute_group_stem_direction(&group.notes, staff_middle_y)
-            };
-
-            let notehead_width = stems::Stem::NOTEHEAD_WIDTH;
-
-            // === PHASE 1: Compute initial stems and beam line ===
-            // Create initial stems (may not reach the beam yet)
-            // Adjust note Y for visual rendering: pitch_to_y subtracts 0.5 staff spaces
-            // to compensate for dominant-baseline:middle in SVG text rendering. But stems
-            // are SVG <line> elements (not text), so they need the visual center Y.
-            let visual_y_offset = 0.5 * units_per_space;
-            let mut initial_stems: Vec<stems::Stem> = Vec::new();
-            let min_length = stems::Stem::MIN_BEAMED_STEM_LENGTH;
-            for note in &group.notes {
-                let visual_y = note.y + visual_y_offset;
-                let mut stem =
-                    stems::create_stem(note.x, visual_y, group_direction, notehead_width);
-
-                // For beamed notes, use MIN_BEAMED_STEM_LENGTH for initial beam
-                // positioning. STEM_LENGTH (3.5 spaces) is for un-beamed notes;
-                // using it here pushes the beam too far from noteheads, causing
-                // excessively long stems on notes far from the beam side.
-                // The beam offset logic below will extend stems outward as needed.
-                match group_direction {
-                    stems::StemDirection::Up => {
-                        stem.y_end = stem.y_start - min_length;
-                    }
-                    stems::StemDirection::Down => {
-                        stem.y_end = stem.y_start + min_length;
-                    }
-                }
-                initial_stems.push(stem);
-            }
-
-            // Compute the beam line from first and last stem endpoints
-            let first_stem_end = initial_stems[0].y_end;
-            let last_stem_end = initial_stems.last().unwrap().y_end;
-            let first_stem_x = initial_stems[0].x;
-            let last_stem_x = initial_stems.last().unwrap().x;
-            let dx = last_stem_x - first_stem_x;
-            let beam_slope = if dx.abs() > 0.001 {
-                (last_stem_end - first_stem_end) / dx
-            } else {
-                0.0
-            };
-
-            // Clamp beam slope
-            let max_slope_units = beams::Beam::MAX_SLOPE * units_per_space;
-            let max_slope_per_unit = if dx.abs() > 0.001 {
-                max_slope_units / dx.abs()
-            } else {
-                0.0
-            };
-            let clamped_slope = beam_slope.clamp(-max_slope_per_unit, max_slope_per_unit);
-
-            // Compute the beam Y at each stem's X position
-            // Then find the "outermost" beam position that ensures ALL stems are long enough
-            // For stem-up: beam is above (smaller Y), so we need the MINIMUM y_end
-            // For stem-down: beam is below (larger Y), so we need the MAXIMUM y_end
-            let mut beam_y_at_stems: Vec<f32> = Vec::new();
-            for stem in &initial_stems {
-                let beam_y = first_stem_end + clamped_slope * (stem.x - first_stem_x);
-                beam_y_at_stems.push(beam_y);
-            }
-
-            // Check if any stem is too short to reach the beam line and adjust
-            // The beam must be positioned so ALL stems can reach it
-            let mut beam_offset = 0.0f32;
-            for (i, stem) in initial_stems.iter().enumerate() {
-                let beam_y = beam_y_at_stems[i];
-                let min_length = stems::Stem::MIN_BEAMED_STEM_LENGTH;
-                match group_direction {
-                    stems::StemDirection::Up => {
-                        // Stem goes up (negative Y). beam_y should be <= stem.y_start - min_length
-                        let required_beam_y = stem.y_start - min_length;
-                        if beam_y > required_beam_y {
-                            // Beam needs to move further up (more negative)
-                            let needed_offset = required_beam_y - beam_y;
-                            beam_offset = beam_offset.min(needed_offset);
-                        }
-                    }
-                    stems::StemDirection::Down => {
-                        // Stem goes down (positive Y). beam_y should be >= stem.y_start + min_length
-                        let required_beam_y = stem.y_start + min_length;
-                        if beam_y < required_beam_y {
-                            // Beam needs to move further down (more positive)
-                            let needed_offset = required_beam_y - beam_y;
-                            beam_offset = beam_offset.max(needed_offset);
-                        }
-                    }
-                }
-            }
-
-            // === PHASE 2: Extend ALL stems to reach the beam line ===
-            let mut stem_end_ys = Vec::new();
-            let mut stem_xs = Vec::new();
-
-            for (i, stem) in initial_stems.iter().enumerate() {
-                let adjusted_beam_y = beam_y_at_stems[i] + beam_offset;
-                // Extend stem to reach the beam
-                let final_stem_end = adjusted_beam_y;
-
-                // Create stem glyph with adjusted length
-                let stem_top_y = stem.y_start.min(final_stem_end);
-                let stem_height = (final_stem_end - stem.y_start).abs();
-
-                let stem_glyph = Glyph {
-                    codepoint: '\u{0000}'.to_string(),
-                    position: Point {
-                        x: stem.x,
-                        y: stem_top_y,
-                    },
-                    bounding_box: BoundingBox {
-                        x: stem.x - (stem.thickness / 2.0),
-                        y: stem_top_y,
-                        width: stem.thickness,
-                        height: stem_height,
-                    },
-                    source_reference: SourceReference {
-                        instrument_id: instrument_id.to_string(),
-                        staff_index,
-                        voice_index,
-                        event_index: group.notes[i].event_index,
-                    },
-                    font_size: None,
-                };
-                stem_glyphs.push(stem_glyph);
-                stem_end_ys.push(final_stem_end);
-                stem_xs.push(stem.x);
-            }
-
-            // === PHASE 3: Create beam at the adjusted position ===
-            let beamable_with_stems: Vec<beams::BeamableNote> = group
-                .notes
-                .iter()
-                .zip(stem_end_ys.iter().zip(stem_xs.iter()))
-                .map(|(n, (&stem_end_y, &stem_x))| beams::BeamableNote {
-                    x: stem_x,
-                    y: n.y,
-                    stem_end_y,
-                    tick: n.tick,
-                    duration_ticks: n.duration_ticks,
-                    beam_levels: n.beam_levels,
-                    beam_types: n.beam_types.clone(),
-                    event_index: n.event_index,
-                })
-                .collect();
-
-            // Create beam — slope is already embedded in the stem_end_y values
-            let slope = clamped_slope;
-            if let Some(beam) = beams::create_beam(&beamable_with_stems, slope) {
-                // Encode beam as special glyph (U+0001)
-                // For sloped beam rendering, we encode:
-                //   position.x/y = left-side beam start (x_start, y_start)
-                //   bounding_box.x = x_start
-                //   bounding_box.y = y_end (right-side Y for slope reconstruction)
-                //   bounding_box.width = horizontal span (x_end - x_start)
-                //   bounding_box.height = beam thickness
-                let beam_glyph = Glyph {
-                    codepoint: '\u{0001}'.to_string(),
-                    position: Point {
-                        x: beam.x_start,
-                        y: beam.y_start,
-                    },
-                    bounding_box: BoundingBox {
-                        x: beam.x_start,
-                        y: beam.y_end, // Right-side Y for slope
-                        width: beam.x_end - beam.x_start,
-                        height: beam.thickness,
-                    },
-                    source_reference: SourceReference {
-                        instrument_id: instrument_id.to_string(),
-                        staff_index,
-                        voice_index,
-                        // Beams use the first note's event_index — they visually span
-                        // the whole group but clicking a beam selects the group's first note.
-                        event_index: group.notes.first().map_or(0, |n| n.event_index),
-                    },
-                    font_size: None,
-                };
-                all_glyphs.push(beam_glyph);
-            }
-
-            // T027: Create secondary/tertiary beam levels for 16th, 32nd notes, etc.
-            // Use beamable_with_stems (with correct stem_end_y) instead of original group
-            let stem_direction_up = group_direction == stems::StemDirection::Up;
-            let updated_group = beams::BeamGroup {
-                notes: beamable_with_stems,
-                beam_count: group.beam_count,
-            };
-            let multi_beams =
-                beams::create_multi_level_beams(&updated_group, slope, stem_direction_up);
-            for beam in multi_beams {
-                let beam_glyph = Glyph {
-                    codepoint: '\u{0001}'.to_string(),
-                    position: Point {
-                        x: beam.x_start,
-                        y: beam.y_start,
-                    },
-                    bounding_box: BoundingBox {
-                        x: beam.x_start,
-                        y: beam.y_end, // Right-side Y for slope
-                        width: beam.x_end - beam.x_start,
-                        height: beam.thickness,
-                    },
-                    source_reference: SourceReference {
-                        instrument_id: instrument_id.to_string(),
-                        staff_index,
-                        voice_index,
-                        // Same as primary beam: use first note's event_index.
-                        event_index: updated_group.notes.first().map_or(0, |n| n.event_index),
-                    },
-                    font_size: None,
-                };
-                all_glyphs.push(beam_glyph);
-            }
-        }
-
-        all_glyphs.extend(stem_glyphs);
-    }
-
-    // T015: Position rests for this staff (all voices combined)
-    let all_staff_rests: Vec<RestLayoutEvent> = staff_data
-        .voices
-        .iter()
-        .flat_map(|v| v.rests.iter().cloned())
-        .collect();
-
-    if !all_staff_rests.is_empty() {
-        let multi_voice = staff_data.voices.len() > 1;
-        let rest_glyphs = positioner::position_rests_for_staff(
-            &all_staff_rests,
-            tick_range.start_tick,
-            tick_range.end_tick,
-            note_positions,
-            staff_data.time_numerator,
-            staff_data.time_denominator,
-            multi_voice,
-            units_per_space,
-            staff_vertical_offset,
-            left_margin,
-            instrument_id,
-            staff_index,
-            measure_x_bounds,
-            pickup_ticks,
-        );
-        all_glyphs.extend(rest_glyphs);
-    }
-
-    all_glyphs
-}
-
-/// Compute the vertical extent of notes in a staff for a given tick range.
-///
-/// Returns (min_y, max_y) relative to the staff origin (top line = 0).
-/// min_y ≤ 0 means notes extend above the staff; max_y ≥ 4*ups means notes
-/// extend below the staff. The baseline extent is always [0, 4*ups] (the
-/// five staff lines themselves).
-fn compute_staff_note_extents(
-    staff_data: &StaffData,
-    tick_range: &TickRange,
-    units_per_space: f32,
-) -> (f32, f32) {
-    let mut min_y = 0.0_f32; // top staff line
-    let mut max_y = 4.0 * units_per_space; // bottom staff line
-    let middle_y = 2.0 * units_per_space; // middle staff line
-
-    for voice in &staff_data.voices {
-        for note in &voice.notes {
-            if note.start_tick >= tick_range.start_tick && note.start_tick < tick_range.end_tick {
-                let active_clef = staff_data.get_clef_at_tick(note.start_tick);
-                let y = positioner::pitch_to_y_with_spelling(
-                    note.pitch,
-                    active_clef,
-                    units_per_space,
-                    note.spelling,
-                );
-                if y < min_y {
-                    min_y = y;
-                }
-                if y > max_y {
-                    max_y = y;
-                }
-                // Include stem extent: notes above the middle line have stems going
-                // DOWN (increasing y); notes on/below the middle line have stems
-                // going UP (decreasing y).  Both can push into the inter-staff gap.
-                if y < middle_y {
-                    let stem_tip = y + stems::Stem::STEM_LENGTH;
-                    if stem_tip > max_y {
-                        max_y = stem_tip;
-                    }
-                } else {
-                    let stem_tip = y - stems::Stem::STEM_LENGTH;
-                    if stem_tip < min_y {
-                        min_y = stem_tip;
-                    }
-                }
-            }
-        }
-    }
-
-    (min_y, max_y)
-}
-
-/// Shift a dot y-coordinate so it sits in a staff space (not on a line).
-///
-/// Staff lines are at `staff_offset + n * units_per_space` for n=0..4.
-/// If y is within a small tolerance of a line, nudge it up by half a space.
-fn shift_dot_to_space(y: f32, staff_offset: f32, units_per_space: f32) -> f32 {
-    let relative = y - staff_offset;
-    let line_index = (relative / units_per_space).round();
-    let on_line = (relative - line_index * units_per_space).abs() < 0.15 * units_per_space;
-    if on_line && (0.0..=4.0).contains(&line_index) {
-        // Shift up (visually) by half a space — move towards the nearest space above the line
-        y - 0.5 * units_per_space
-    } else {
-        y
-    }
-}
-
-/// Create staff lines for a single staff
-fn create_staff_lines(
-    staff_vertical_offset: f32,
-    system_width: f32,
-    units_per_space: f32,
-) -> [StaffLine; 5] {
-    // Create 5 evenly spaced lines (1 staff space apart)
-    // A notehead fills exactly this gap, matching standard engraving
-    let mut lines = Vec::new();
-    for line_index in 0..5 {
-        let y_position = staff_vertical_offset + (line_index as f32 * units_per_space);
-        lines.push(StaffLine {
-            y_position,
-            start_x: 0.0,
-            end_x: system_width,
-        });
-    }
-
-    // Convert Vec to array (guaranteed to have exactly 5 elements)
-    lines.try_into().unwrap()
-}
-
-/// Create bar lines for a single staff at measure boundaries
-///
-/// # Arguments
-/// * `measure_infos` - All measures with their tick ranges and widths
-/// * `tick_range` - The system's tick range
-/// * `staff_index` - Index of staff for vertical positioning
-/// * `left_margin` - Left margin where music content starts
-/// * `system_width` - Total width of the system
-/// * `units_per_space` - Scaling factor
-/// * `system_y_position` - Vertical position of the system
-///
-/// # Returns
-/// Vector of bar lines positioned at measure boundaries
-#[allow(clippy::too_many_arguments)]
-fn create_bar_lines(
-    measure_infos: &[breaker::MeasureInfo],
-    tick_range: &TickRange,
-    staff_vertical_offset: f32,
-    left_margin: f32,
-    _system_width: f32,
-    units_per_space: f32,
-    note_positions: &HashMap<u32, f32>,
-    measure_x_bounds: &HashMap<u32, (f32, f32)>,
-) -> Vec<BarLine> {
-    let mut bar_lines = Vec::new();
-
-    // Y positions for top and bottom staff lines
-    let y_start = staff_vertical_offset; // Top line (line 0)
-    let y_end = staff_vertical_offset + (4.0 * units_per_space); // Bottom line (line 4)
-
-    // Find measures that overlap with this system's tick range
-    let measures_in_system: Vec<&breaker::MeasureInfo> = measure_infos
-        .iter()
-        .filter(|m| m.start_tick < tick_range.end_tick && m.end_tick > tick_range.start_tick)
-        .collect();
-
-    if measures_in_system.is_empty() {
-        return bar_lines;
-    }
-
-    // Add bar lines at the end of each measure.
-    // Use the measure's end boundary from measure_x_bounds (which are already
-    // justified/aligned) so barlines appear at the correct position between
-    // the last note of one measure and the first note of the next.
-    for measure in measures_in_system.iter() {
-        // Skip barline if measure ends at or beyond the system boundary.
-        // The system-end barline post-processing step handles the closing
-        // barline at the justified edge, so we must not create a duplicate
-        // that would leave an empty gap before the system edge.
-        if measure.end_tick >= tick_range.end_tick {
-            continue;
-        }
-
-        // Place barline at the midpoint between the last note of this measure
-        // and the first note of the next, ensuring equal visual clearance on
-        // both sides.  Fall back to a fixed offset when a side has no notes
-        // (empty measures or last measure in the system).
-        let last_x_in_measure = note_positions
-            .iter()
-            .filter(|(tick, _)| **tick >= measure.start_tick && **tick < measure.end_tick)
-            .max_by_key(|(tick, _)| *tick)
-            .map(|(_, &x)| x);
-        let first_x_in_next = note_positions
-            .iter()
-            .filter(|(tick, _)| **tick >= measure.end_tick && **tick < tick_range.end_tick)
-            .min_by_key(|(tick, _)| *tick)
-            .map(|(_, &x)| x);
-        let barline_x = match (last_x_in_measure, first_x_in_next) {
-            (Some(last_x), Some(next_x)) if next_x > last_x => (last_x + next_x) * 0.5,
-            (Some(last_x), _) => last_x + 30.0,
-            _ => measure_x_bounds
-                .get(&measure.start_tick)
-                .map(|(_, end_x)| *end_x)
-                .unwrap_or(left_margin + 30.0),
-        };
-
-        // Determine bar line type
-        let bar_type = {
-            // Check repeat flags for this measure and the one that starts next
-            let next_start_repeat = measure_infos
-                .iter()
-                .find(|m| m.start_tick == measure.end_tick)
-                .map(|m| m.start_repeat)
-                .unwrap_or(false);
-
-            if measure.end_repeat && next_start_repeat {
-                BarLineType::RepeatBoth
-            } else if measure.end_repeat {
-                BarLineType::RepeatEnd
-            } else if next_start_repeat {
-                BarLineType::RepeatStart
-            } else if measure.end_tick == tick_range.end_tick {
-                // Check if this is the very last measure in the entire score
-                let is_last_measure =
-                    measure_infos.last().map(|m| m.end_tick) == Some(measure.end_tick);
-                if is_last_measure {
-                    BarLineType::Final
-                } else {
-                    BarLineType::Single
-                }
-            } else {
-                BarLineType::Single
-            }
-        };
-
-        // Create bar line segments with explicit geometry (Principle VI: Layout Engine Authority)
-        let segments = create_bar_line_segments(barline_x, y_start, y_end, &bar_type);
-        let dots = compute_repeat_dots(barline_x, y_start, units_per_space, &bar_type);
-
-        bar_lines.push(BarLine {
-            segments,
-            bar_type,
-            dots,
-        });
-    }
-
-    bar_lines
-}
-
-/// Creates bar line segments with explicit geometry for each line
-/// This ensures the renderer doesn't calculate positions (Principle VI compliance)
-fn create_bar_line_segments(
-    x_position: f32,
-    y_start: f32,
-    y_end: f32,
-    bar_type: &BarLineType,
-) -> Vec<BarLineSegment> {
-    const THIN_WIDTH: f32 = 1.5;
-    const THICK_WIDTH: f32 = 4.0;
-    const DOUBLE_SPACING: f32 = 4.0; // Space between double bar lines
-    const FINAL_SPACING: f32 = 4.0; // Space between thin and thick in final bar
-
-    match bar_type {
-        BarLineType::Single => {
-            vec![BarLineSegment {
-                x_position,
-                y_start,
-                y_end,
-                stroke_width: THIN_WIDTH,
-            }]
-        }
-        BarLineType::Double => {
-            vec![
-                BarLineSegment {
-                    x_position: x_position - DOUBLE_SPACING / 2.0,
-                    y_start,
-                    y_end,
-                    stroke_width: THIN_WIDTH,
-                },
-                BarLineSegment {
-                    x_position: x_position + DOUBLE_SPACING / 2.0,
-                    y_start,
-                    y_end,
-                    stroke_width: THIN_WIDTH,
-                },
-            ]
-        }
-        BarLineType::Final => {
-            vec![
-                BarLineSegment {
-                    x_position: x_position - FINAL_SPACING - THICK_WIDTH / 2.0,
-                    y_start,
-                    y_end,
-                    stroke_width: THIN_WIDTH,
-                },
-                BarLineSegment {
-                    x_position: x_position + THICK_WIDTH / 2.0,
-                    y_start,
-                    y_end,
-                    stroke_width: THICK_WIDTH,
-                },
-            ]
-        }
-        BarLineType::RepeatEnd => {
-            // Same visual as Final: thin bar on left, thick bar on right
-            vec![
-                BarLineSegment {
-                    x_position: x_position - FINAL_SPACING - THICK_WIDTH / 2.0,
-                    y_start,
-                    y_end,
-                    stroke_width: THIN_WIDTH,
-                },
-                BarLineSegment {
-                    x_position: x_position + THICK_WIDTH / 2.0,
-                    y_start,
-                    y_end,
-                    stroke_width: THICK_WIDTH,
-                },
-            ]
-        }
-        BarLineType::RepeatStart => {
-            // Mirror of Final: thick bar on left, thin bar on right
-            vec![
-                BarLineSegment {
-                    x_position: x_position - THICK_WIDTH / 2.0,
-                    y_start,
-                    y_end,
-                    stroke_width: THICK_WIDTH,
-                },
-                BarLineSegment {
-                    x_position: x_position + FINAL_SPACING + THICK_WIDTH / 2.0,
-                    y_start,
-                    y_end,
-                    stroke_width: THIN_WIDTH,
-                },
-            ]
-        }
-        BarLineType::RepeatBoth => {
-            // Combined end-repeat and start-repeat: thin-thick-thick-thin
-            vec![
-                BarLineSegment {
-                    x_position: x_position - FINAL_SPACING - THICK_WIDTH,
-                    y_start,
-                    y_end,
-                    stroke_width: THIN_WIDTH,
-                },
-                BarLineSegment {
-                    x_position: x_position - THICK_WIDTH / 2.0,
-                    y_start,
-                    y_end,
-                    stroke_width: THICK_WIDTH,
-                },
-                BarLineSegment {
-                    x_position: x_position + THICK_WIDTH / 2.0,
-                    y_start,
-                    y_end,
-                    stroke_width: THICK_WIDTH,
-                },
-                BarLineSegment {
-                    x_position: x_position + FINAL_SPACING + THICK_WIDTH,
-                    y_start,
-                    y_end,
-                    stroke_width: THIN_WIDTH,
-                },
-            ]
-        }
-    }
-}
-
-/// Compute repeat dot positions for repeat barline types (Principle VI: Layout Engine Authority)
-///
-/// All dot coordinates are computed here in Rust. TypeScript renders them verbatim.
-fn compute_repeat_dots(
-    x_position: f32,
-    y_start: f32,
-    units_per_space: f32,
-    bar_type: &BarLineType,
-) -> Vec<RepeatDotPosition> {
-    const THICK_WIDTH: f32 = 4.0;
-    let dot_radius = 0.25 * units_per_space;
-    let x_offset = 0.6 * units_per_space;
-    let dot_y0 = y_start + 1.0 * units_per_space;
-    let dot_y1 = y_start + 3.0 * units_per_space;
-
-    match bar_type {
-        BarLineType::RepeatEnd => {
-            // Thick bar center at x_position + THICK_WIDTH/2.0; dots to its left
-            let dot_x = x_position + THICK_WIDTH / 2.0 - x_offset;
-            vec![
-                RepeatDotPosition {
-                    x: dot_x,
-                    y: dot_y0,
-                    radius: dot_radius,
-                },
-                RepeatDotPosition {
-                    x: dot_x,
-                    y: dot_y1,
-                    radius: dot_radius,
-                },
-            ]
-        }
-        BarLineType::RepeatStart => {
-            // Thick bar center at x_position - THICK_WIDTH/2.0; dots to its right
-            let dot_x = x_position - THICK_WIDTH / 2.0 + x_offset;
-            vec![
-                RepeatDotPosition {
-                    x: dot_x,
-                    y: dot_y0,
-                    radius: dot_radius,
-                },
-                RepeatDotPosition {
-                    x: dot_x,
-                    y: dot_y1,
-                    radius: dot_radius,
-                },
-            ]
-        }
-        BarLineType::RepeatBoth => {
-            // Left thick bar (end-repeat) at x_position - THICK_WIDTH/2.0
-            // Right thick bar (start-repeat) at x_position + THICK_WIDTH/2.0
-            let dot_x_left = x_position - THICK_WIDTH / 2.0 - x_offset;
-            let dot_x_right = x_position + THICK_WIDTH / 2.0 + x_offset;
-            vec![
-                RepeatDotPosition {
-                    x: dot_x_left,
-                    y: dot_y0,
-                    radius: dot_radius,
-                },
-                RepeatDotPosition {
-                    x: dot_x_left,
-                    y: dot_y1,
-                    radius: dot_radius,
-                },
-                RepeatDotPosition {
-                    x: dot_x_right,
-                    y: dot_y0,
-                    radius: dot_radius,
-                },
-                RepeatDotPosition {
-                    x: dot_x_right,
-                    y: dot_y1,
-                    radius: dot_radius,
-                },
-            ]
-        }
-        _ => vec![],
-    }
-}
-
-/// Creates bracket/brace glyph with calculated geometry for multi-staff groups
-fn create_bracket_glyph(
-    staves: &[Staff],
-    bracket_type: &BracketType,
-    _config: &LayoutConfig,
-) -> BracketGlyph {
-    let first_staff = &staves[0];
-    let last_staff = &staves[staves.len() - 1];
-
-    // Span from top of first staff to bottom of last staff
-    let top_y = first_staff.staff_lines[0].y_position;
-    let bottom_y = last_staff.staff_lines[4].y_position;
-    let height = bottom_y - top_y;
-    let _center_y = (top_y + bottom_y) / 2.0;
-
-    // Scale glyph to match actual bracket height
-    const BRACE_NATURAL_HEIGHT: f32 = 320.0; // SMuFL brace U+E000 at fontSize 80
-    let scale_y = height / BRACE_NATURAL_HEIGHT;
-
-    let codepoint = match bracket_type {
-        BracketType::Brace => "\u{E000}".to_string(),
-        BracketType::Bracket => "\u{E002}".to_string(),
-        BracketType::None => String::new(),
-    };
-
-    let x_position = 5.0; // Left margin
-
-    BracketGlyph {
-        codepoint,
-        x: x_position,
-        // Feature 027 (T034): Use top_y (top anchor) instead of center_y.
-        // SMuFL brace U+E000 is rendered with dominant-baseline="hanging" in the
-        // frontend, anchoring the glyph top to this y coordinate. Using center_y
-        // caused the brace to appear shifted down by half its height.
-        y: top_y,
-        scale_y,
-        bounding_box: BoundingBox {
-            x: x_position - 5.0,
-            y: top_y,
-            width: 20.0,
-            height,
-        },
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    /// T011: Unit test for create_staff_lines() verifying 5 lines with 20-unit spacing
-    #[test]
-    fn test_create_staff_lines_spacing() {
-        let units_per_space = 20.0;
-        let system_width = 1200.0;
-
-        let lines = create_staff_lines(0.0, system_width, units_per_space);
-
-        // Verify exactly 5 lines
-        assert_eq!(lines.len(), 5, "Should have exactly 5 staff lines");
-
-        // Verify y-positions with 20-unit spacing (1 * units_per_space)
-        assert_eq!(lines[0].y_position, 0.0, "Line 0 should be at y=0");
-        assert_eq!(lines[1].y_position, 20.0, "Line 1 should be at y=20");
-        assert_eq!(lines[2].y_position, 40.0, "Line 2 should be at y=40");
-        assert_eq!(lines[3].y_position, 60.0, "Line 3 should be at y=60");
-        assert_eq!(lines[4].y_position, 80.0, "Line 4 should be at y=80");
-
-        // Verify all lines span full system width
-        for (i, line) in lines.iter().enumerate() {
-            assert_eq!(line.start_x, 0.0, "Line {} should start at x=0", i);
-            assert_eq!(
-                line.end_x, system_width,
-                "Line {} should end at system_width",
-                i
-            );
-        }
-    }
-
-    /// T011: Unit test for multi-staff layout with correct vertical offsetting
-    #[test]
-    fn test_create_staff_lines_multi_staff() {
-        let units_per_space = 20.0;
-        let system_width = 1200.0;
-
-        // First staff at vertical offset 0
-        let staff_0 = create_staff_lines(0.0, system_width, units_per_space);
-        assert_eq!(staff_0[0].y_position, 0.0);
-        assert_eq!(staff_0[4].y_position, 80.0);
-
-        // Second staff - offset by 20 staff spaces (400 units)
-        let expected_offset = 20.0 * units_per_space; // 400 units
-        let staff_1 = create_staff_lines(expected_offset, system_width, units_per_space);
-        assert_eq!(staff_1[0].y_position, expected_offset);
-        assert_eq!(staff_1[4].y_position, expected_offset + 80.0);
-    }
-
-    /// T011: Unit test for different units_per_space values
-    #[test]
-    fn test_create_staff_lines_scale_independence() {
-        let system_width = 1200.0;
-
-        // Test with different scale (units_per_space = 10)
-        let lines_scale_10 = create_staff_lines(0.0, system_width, 10.0);
-        assert_eq!(lines_scale_10[0].y_position, 0.0);
-        assert_eq!(lines_scale_10[1].y_position, 10.0); // 1 * 10
-        assert_eq!(lines_scale_10[4].y_position, 40.0); // 4 * 10
-
-        // Test with different scale (units_per_space = 25)
-        let lines_scale_25 = create_staff_lines(0.0, system_width, 25.0);
-        assert_eq!(lines_scale_25[0].y_position, 0.0);
-        assert_eq!(lines_scale_25[1].y_position, 25.0); // 1 * 25
-        assert_eq!(lines_scale_25[4].y_position, 100.0); // 4 * 25
-    }
 
     /// T029: Integration test for structural glyph generation
     #[test]
@@ -3745,866 +923,6 @@ mod tests {
         assert!(
             staff_group.bracket_glyph.is_some(),
             "Piano staff group should have bracket glyph"
-        );
-    }
-
-    /// T064: Unit test for brace/bracket positioning and vertical scaling
-    #[test]
-    fn test_create_bracket_glyph_brace() {
-        let config = LayoutConfig::default();
-
-        // Create two dummy staves at different vertical positions
-        let staff_0_lines = create_staff_lines(0.0, 1200.0, config.units_per_space);
-        let staff_1_offset = 14.0 * config.units_per_space;
-        let staff_1_lines = create_staff_lines(staff_1_offset, 1200.0, config.units_per_space);
-
-        let staff_0 = Staff {
-            staff_lines: staff_0_lines,
-            glyph_runs: vec![],
-            structural_glyphs: vec![],
-            bar_lines: vec![],
-            ledger_lines: vec![],
-            notation_dots: vec![],
-            tie_arcs: vec![],
-            slur_arcs: vec![],
-        };
-
-        let staff_1 = Staff {
-            staff_lines: staff_1_lines,
-            glyph_runs: vec![],
-            structural_glyphs: vec![],
-            bar_lines: vec![],
-            ledger_lines: vec![],
-            notation_dots: vec![],
-            tie_arcs: vec![],
-            slur_arcs: vec![],
-        };
-
-        let staves = vec![staff_0, staff_1];
-        let bracket_type = BracketType::Brace;
-
-        let bracket_glyph = create_bracket_glyph(&staves, &bracket_type, &config);
-
-        // Verify brace codepoint
-        assert_eq!(
-            bracket_glyph.codepoint, "\u{E000}",
-            "Brace should use SMuFL codepoint U+E000"
-        );
-
-        // Verify x position (left margin)
-        assert_eq!(bracket_glyph.x, 5.0, "Brace should be at x=5");
-
-        // Verify vertical scaling is applied
-        assert!(
-            bracket_glyph.scale_y > 0.0,
-            "Brace should have positive vertical scale"
-        );
-
-        // Verify bounding box spans both staves
-        let first_staff_top = staves[0].staff_lines[0].y_position;
-        let last_staff_bottom = staves[1].staff_lines[4].y_position;
-
-        // Brace should extend to cover both staves
-        assert!(
-            bracket_glyph.bounding_box.y <= first_staff_top,
-            "Brace bounding box should start at or above first staff"
-        );
-        assert!(
-            bracket_glyph.bounding_box.y + bracket_glyph.bounding_box.height >= last_staff_bottom,
-            "Brace bounding box should extend to or below last staff"
-        );
-    }
-
-    /// T064: Unit test for square bracket positioning (ensemble scores)
-    #[test]
-    fn test_create_bracket_glyph_bracket() {
-        let config = LayoutConfig::default();
-
-        let staff_0_lines = create_staff_lines(0.0, 1200.0, config.units_per_space);
-        let staff_1_offset = 14.0 * config.units_per_space;
-        let staff_1_lines = create_staff_lines(staff_1_offset, 1200.0, config.units_per_space);
-
-        let staff_0 = Staff {
-            staff_lines: staff_0_lines,
-            glyph_runs: vec![],
-            structural_glyphs: vec![],
-            bar_lines: vec![],
-            ledger_lines: vec![],
-            notation_dots: vec![],
-            tie_arcs: vec![],
-            slur_arcs: vec![],
-        };
-
-        let staff_1 = Staff {
-            staff_lines: staff_1_lines,
-            glyph_runs: vec![],
-            structural_glyphs: vec![],
-            bar_lines: vec![],
-            ledger_lines: vec![],
-            notation_dots: vec![],
-            tie_arcs: vec![],
-            slur_arcs: vec![],
-        };
-
-        let staves = vec![staff_0, staff_1];
-        let bracket_type = BracketType::Bracket;
-
-        let bracket_glyph = create_bracket_glyph(&staves, &bracket_type, &config);
-
-        // Verify bracket codepoint
-        assert_eq!(
-            bracket_glyph.codepoint, "\u{E002}",
-            "Bracket should use SMuFL codepoint U+E002"
-        );
-    }
-
-    /// T074: Test notes on both staves render correctly relative to their respective staff lines
-    #[test]
-    fn test_notes_on_multi_staff() {
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [
-                    {
-                        "clef": "Treble",
-                        "time_signature": { "numerator": 4, "denominator": 4 },
-                        "key_signature": { "sharps": 0 },
-                        "voices": [{
-                            "notes": [
-                                { "pitch": 72, "tick": 0, "duration": 960 }
-                            ]
-                        }]
-                    },
-                    {
-                        "clef": "Bass",
-                        "time_signature": { "numerator": 4, "denominator": 4 },
-                        "key_signature": { "sharps": 0 },
-                        "voices": [{
-                            "notes": [
-                                { "pitch": 48, "tick": 0, "duration": 960 }
-                            ]
-                        }]
-                    }
-                ]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        let system = &layout.systems[0];
-        let staff_group = &system.staff_groups[0];
-
-        // Both staves should have glyph runs (noteheads)
-        assert!(
-            !staff_group.staves[0].glyph_runs.is_empty(),
-            "Treble staff should have glyphs"
-        );
-        assert!(
-            !staff_group.staves[1].glyph_runs.is_empty(),
-            "Bass staff should have glyphs"
-        );
-
-        // Verify treble staff note is positioned relative to treble staff lines
-        let treble_line_0 = staff_group.staves[0].staff_lines[0].y_position;
-        let treble_glyph = &staff_group.staves[0].glyph_runs[0].glyphs[0];
-        assert!(
-            treble_glyph.position.y >= treble_line_0 - 100.0,
-            "Treble note should be near treble staff"
-        );
-
-        // Verify bass staff note is positioned relative to bass staff lines
-        let bass_line_0 = staff_group.staves[1].staff_lines[0].y_position;
-        let bass_glyph = &staff_group.staves[1].glyph_runs[0].glyphs[0];
-        assert!(
-            bass_glyph.position.y >= bass_line_0 - 100.0,
-            "Bass note should be near bass staff"
-        );
-
-        // Verify bass staff is below treble staff
-        assert!(
-            bass_line_0 > treble_line_0,
-            "Bass staff should be below treble staff"
-        );
-    }
-
-    /// T020: Integration test — 4 eighth notes with beam info produce correct glyphs
-    #[test]
-    fn test_four_beamed_eighths_produce_noteheads_stems_beam() {
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [{
-                    "clef": "Treble",
-                    "time_signature": { "numerator": 4, "denominator": 4 },
-                    "key_signature": { "sharps": 0 },
-                    "voices": [{
-                        "notes": [
-                            { "pitch": 72, "tick": 0, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "Begin"}] },
-                            { "pitch": 74, "tick": 480, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "Continue"}] },
-                            { "pitch": 76, "tick": 960, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "Continue"}] },
-                            { "pitch": 77, "tick": 1440, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "End"}] }
-                        ]
-                    }]
-                }]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        // Collect all glyphs across all runs
-        let staff = &layout.systems[0].staff_groups[0].staves[0];
-        let all_glyphs: Vec<_> = staff
-            .glyph_runs
-            .iter()
-            .flat_map(|run| run.glyphs.iter())
-            .collect();
-
-        // Count glyph types
-        let noteheads = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{E0A4}")
-            .count();
-        let stems = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0000}")
-            .count();
-        let beams = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0001}")
-            .count();
-        let combined_eighth = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{E1D7}")
-            .count();
-
-        assert_eq!(
-            noteheads, 4,
-            "Should have 4 bare noteheadBlack glyphs (U+E0A4)"
-        );
-        assert_eq!(stems, 4, "Should have 4 stem glyphs (U+0000)");
-        assert_eq!(beams, 1, "Should have 1 beam glyph (U+0001)");
-        assert_eq!(
-            combined_eighth, 0,
-            "Should have 0 combined eighth note glyphs (U+E1D7)"
-        );
-    }
-
-    /// T021: Integration test — mixed quarters and eighths produce correct glyphs
-    #[test]
-    fn test_mixed_quarters_and_beamed_eighths() {
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [{
-                    "clef": "Treble",
-                    "time_signature": { "numerator": 4, "denominator": 4 },
-                    "key_signature": { "sharps": 0 },
-                    "voices": [{
-                        "notes": [
-                            { "pitch": 72, "tick": 0, "duration": 960 },
-                            { "pitch": 74, "tick": 960, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "Begin"}] },
-                            { "pitch": 76, "tick": 1440, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "End"}] },
-                            { "pitch": 77, "tick": 1920, "duration": 960 }
-                        ]
-                    }]
-                }]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        let staff = &layout.systems[0].staff_groups[0].staves[0];
-        let all_glyphs: Vec<_> = staff
-            .glyph_runs
-            .iter()
-            .flat_map(|run| run.glyphs.iter())
-            .collect();
-
-        // Count glyph types (check both Up and Down variants for direction-aware glyphs)
-        let combined_quarter = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{E1D5}" || g.codepoint == "\u{E1D6}")
-            .count();
-        let bare_noteheads = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{E0A4}")
-            .count();
-        let stems = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0000}")
-            .count();
-        let beams = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0001}")
-            .count();
-        let combined_eighth = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{E1D7}" || g.codepoint == "\u{E1D8}")
-            .count();
-
-        assert_eq!(
-            combined_quarter, 2,
-            "Should have 2 combined quarter note glyphs (U+E1D5 or U+E1D6)"
-        );
-        assert_eq!(
-            bare_noteheads, 2,
-            "Should have 2 bare noteheadBlack for beamed eighths"
-        );
-        assert_eq!(stems, 2, "Should have 2 stem glyphs for beamed eighths");
-        assert_eq!(
-            beams, 1,
-            "Should have 1 beam glyph connecting the 2 eighths"
-        );
-        assert_eq!(combined_eighth, 0, "Should have 0 combined eighth glyphs");
-    }
-
-    /// T028: Integration test — 4 sixteenth notes produce 2 beam levels
-    #[test]
-    fn test_four_sixteenths_two_beam_levels() {
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [{
-                    "clef": "Treble",
-                    "time_signature": { "numerator": 4, "denominator": 4 },
-                    "key_signature": { "sharps": 0 },
-                    "voices": [{
-                        "notes": [
-                            { "pitch": 72, "tick": 0, "duration": 240,
-                              "beams": [{"number": 1, "beam_type": "Begin"}, {"number": 2, "beam_type": "Begin"}] },
-                            { "pitch": 74, "tick": 240, "duration": 240,
-                              "beams": [{"number": 1, "beam_type": "Continue"}, {"number": 2, "beam_type": "Continue"}] },
-                            { "pitch": 76, "tick": 480, "duration": 240,
-                              "beams": [{"number": 1, "beam_type": "Continue"}, {"number": 2, "beam_type": "Continue"}] },
-                            { "pitch": 77, "tick": 720, "duration": 240,
-                              "beams": [{"number": 1, "beam_type": "End"}, {"number": 2, "beam_type": "End"}] }
-                        ]
-                    }]
-                }]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        let staff = &layout.systems[0].staff_groups[0].staves[0];
-        let all_glyphs: Vec<_> = staff
-            .glyph_runs
-            .iter()
-            .flat_map(|run| run.glyphs.iter())
-            .collect();
-
-        let beams = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0001}")
-            .count();
-
-        assert_eq!(
-            beams, 2,
-            "4 sixteenth notes should produce 2 beam glyphs (level 1 + level 2)"
-        );
-    }
-
-    /// T029: Integration test — mixed eighths + sixteenths produce correct beaming
-    #[test]
-    fn test_mixed_eighths_sixteenths_multi_level() {
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [{
-                    "clef": "Treble",
-                    "time_signature": { "numerator": 4, "denominator": 4 },
-                    "key_signature": { "sharps": 0 },
-                    "voices": [{
-                        "notes": [
-                            { "pitch": 72, "tick": 0, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "Begin"}] },
-                            { "pitch": 74, "tick": 480, "duration": 240,
-                              "beams": [{"number": 1, "beam_type": "Continue"}, {"number": 2, "beam_type": "Begin"}] },
-                            { "pitch": 76, "tick": 720, "duration": 240,
-                              "beams": [{"number": 1, "beam_type": "End"}, {"number": 2, "beam_type": "End"}] }
-                        ]
-                    }]
-                }]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        let staff = &layout.systems[0].staff_groups[0].staves[0];
-        let all_glyphs: Vec<_> = staff
-            .glyph_runs
-            .iter()
-            .flat_map(|run| run.glyphs.iter())
-            .collect();
-
-        let beams = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0001}")
-            .count();
-        let noteheads = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{E0A4}")
-            .count();
-
-        assert_eq!(noteheads, 3, "Should have 3 bare noteheads (all beamed)");
-        assert_eq!(
-            beams, 2,
-            "Should have 2 beams (level 1 spans all, level 2 spans sixteenths)"
-        );
-    }
-
-    /// T036: High-pitched beamed group → stems down, beam below noteheads
-    #[test]
-    fn test_stem_direction_high_notes_stems_down() {
-        // All notes above middle line (high pitches like C6, D6, E6, F6)
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [{
-                    "clef": "Treble",
-                    "time_signature": { "numerator": 4, "denominator": 4 },
-                    "key_signature": { "sharps": 0 },
-                    "voices": [{
-                        "notes": [
-                            { "pitch": 84, "tick": 0, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "Begin"}] },
-                            { "pitch": 86, "tick": 480, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "End"}] }
-                        ]
-                    }]
-                }]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        let staff = &layout.systems[0].staff_groups[0].staves[0];
-        let all_glyphs: Vec<_> = staff
-            .glyph_runs
-            .iter()
-            .flat_map(|run| run.glyphs.iter())
-            .collect();
-
-        // Find stem glyphs
-        let stem_glyphs: Vec<_> = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0000}")
-            .collect();
-
-        assert_eq!(stem_glyphs.len(), 2, "Should have 2 stems");
-
-        // For high notes, stems should point down (y_end > y_start in screen coordinates)
-        // Find the noteheads to compare positions
-        let notehead_glyphs: Vec<_> = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{E0A4}")
-            .collect();
-
-        assert_eq!(notehead_glyphs.len(), 2, "Should have 2 noteheads");
-
-        // Stem bounding box y + height should extend below the notehead y
-        // (stems down = extending downward from notehead)
-        for stem in &stem_glyphs {
-            let stem_bottom = stem.bounding_box.y + stem.bounding_box.height;
-            assert!(
-                stem_bottom > stem.position.y,
-                "Stem should extend downward (stem_bottom {} > position.y {})",
-                stem_bottom,
-                stem.position.y
-            );
-        }
-    }
-
-    /// T037: Beamed group spanning both sides → uniform direction
-    #[test]
-    fn test_uniform_stem_direction_mixed_positions() {
-        // Mix of notes: 2 above middle, 1 below → majority above → stems down
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [{
-                    "clef": "Treble",
-                    "time_signature": { "numerator": 4, "denominator": 4 },
-                    "key_signature": { "sharps": 0 },
-                    "voices": [{
-                        "notes": [
-                            { "pitch": 48, "tick": 0, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "Begin"}] },
-                            { "pitch": 48, "tick": 480, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "Continue"}] },
-                            { "pitch": 84, "tick": 960, "duration": 480,
-                              "beams": [{"number": 1, "beam_type": "End"}] }
-                        ]
-                    }]
-                }]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        let staff = &layout.systems[0].staff_groups[0].staves[0];
-        let all_glyphs: Vec<_> = staff
-            .glyph_runs
-            .iter()
-            .flat_map(|run| run.glyphs.iter())
-            .collect();
-
-        let stem_glyphs: Vec<_> = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0000}")
-            .collect();
-
-        assert_eq!(stem_glyphs.len(), 3, "Should have 3 stems");
-
-        // All stems should have a consistent direction (same sign of displacement)
-        // Collect stem displacement directions (positive = down, negative = up)
-        let displacements: Vec<f32> = stem_glyphs.iter().map(|s| s.bounding_box.height).collect();
-
-        // All stems should be non-zero height
-        for d in &displacements {
-            assert!(*d > 0.0, "Stem height should be positive: {}", d);
-        }
-    }
-
-    /// T046: Algorithmic beaming for 4/4 without <beam> elements
-    #[test]
-    fn test_algorithmic_beaming_4_4() {
-        // No beam_info → algorithmic fallback groups by beat
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [{
-                    "clef": "Treble",
-                    "time_signature": { "numerator": 4, "denominator": 4 },
-                    "key_signature": { "sharps": 0 },
-                    "voices": [{
-                        "notes": [
-                            { "pitch": 72, "tick": 0, "duration": 480 },
-                            { "pitch": 74, "tick": 480, "duration": 480 },
-                            { "pitch": 76, "tick": 960, "duration": 480 },
-                            { "pitch": 77, "tick": 1440, "duration": 480 }
-                        ]
-                    }]
-                }]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        let staff = &layout.systems[0].staff_groups[0].staves[0];
-        let all_glyphs: Vec<_> = staff
-            .glyph_runs
-            .iter()
-            .flat_map(|run| run.glyphs.iter())
-            .collect();
-
-        let beams = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0001}")
-            .count();
-        let noteheads = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{E0A4}")
-            .count();
-
-        assert_eq!(noteheads, 4, "All 4 eighths should use bare noteheadBlack");
-        assert_eq!(
-            beams, 2,
-            "Algorithmic beaming in 4/4 should produce 2 beam groups (2 per beat)"
-        );
-    }
-
-    /// T047: Single isolated eighth note uses combined flag glyph
-    #[test]
-    fn test_single_eighth_uses_flag() {
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [{
-                    "clef": "Treble",
-                    "time_signature": { "numerator": 4, "denominator": 4 },
-                    "key_signature": { "sharps": 0 },
-                    "voices": [{
-                        "notes": [
-                            { "pitch": 72, "tick": 0, "duration": 480 },
-                            { "pitch": 74, "tick": 960, "duration": 960 },
-                            { "pitch": 76, "tick": 1920, "duration": 960 }
-                        ]
-                    }]
-                }]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        let staff = &layout.systems[0].staff_groups[0].staves[0];
-        let all_glyphs: Vec<_> = staff
-            .glyph_runs
-            .iter()
-            .flat_map(|run| run.glyphs.iter())
-            .collect();
-
-        let combined_eighth = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{E1D7}" || g.codepoint == "\u{E1D8}")
-            .count();
-        let beams = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0001}")
-            .count();
-
-        // Single eighth note can't form a beam group, so it should use the combined flag glyph
-        assert_eq!(
-            combined_eighth, 1,
-            "Single eighth should use combined flag glyph (U+E1D7 or U+E1D8)"
-        );
-        assert_eq!(beams, 0, "No beams for single eighth note");
-    }
-
-    /// T051: Degenerate single-note beam group falls back to flagged rendering
-    #[test]
-    fn test_degenerate_single_note_group_uses_flag() {
-        // A single eighth note at beat boundary with no partner → no beam, uses flag
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [{
-                    "clef": "Treble",
-                    "time_signature": { "numerator": 4, "denominator": 4 },
-                    "key_signature": { "sharps": 0 },
-                    "voices": [{
-                        "notes": [
-                            { "pitch": 72, "tick": 0, "duration": 480 },
-                            { "pitch": 74, "tick": 480, "duration": 960 }
-                        ]
-                    }]
-                }]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        let staff = &layout.systems[0].staff_groups[0].staves[0];
-        let all_glyphs: Vec<_> = staff
-            .glyph_runs
-            .iter()
-            .flat_map(|run| run.glyphs.iter())
-            .collect();
-
-        let combined_eighth = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{E1D7}" || g.codepoint == "\u{E1D8}")
-            .count();
-        let bare_noteheads = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{E0A4}")
-            .count();
-        let beams = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0001}")
-            .count();
-
-        assert_eq!(
-            combined_eighth, 1,
-            "Single eighth should use combined flag glyph (U+E1D7 or U+E1D8)"
-        );
-        assert_eq!(
-            bare_noteheads, 0,
-            "No bare noteheads when no beam group forms"
-        );
-        assert_eq!(beams, 0, "No beams for degenerate single-note group");
-    }
-
-    /// T052: Beams do not cross bar lines
-    #[test]
-    fn test_beams_do_not_cross_barlines() {
-        // 2 eighths at end of measure 1 + 2 eighths at start of measure 2
-        // Should produce 2 separate beam groups, not 1 spanning the barline
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [{
-                    "clef": "Treble",
-                    "time_signature": { "numerator": 4, "denominator": 4 },
-                    "key_signature": { "sharps": 0 },
-                    "voices": [{
-                        "notes": [
-                            { "pitch": 72, "tick": 2880, "duration": 480 },
-                            { "pitch": 74, "tick": 3360, "duration": 480 },
-                            { "pitch": 76, "tick": 3840, "duration": 480 },
-                            { "pitch": 77, "tick": 4320, "duration": 480 }
-                        ]
-                    }]
-                }]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        let staff = &layout.systems[0].staff_groups[0].staves[0];
-        let all_glyphs: Vec<_> = staff
-            .glyph_runs
-            .iter()
-            .flat_map(|run| run.glyphs.iter())
-            .collect();
-
-        let beams = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0001}")
-            .count();
-
-        assert_eq!(
-            beams, 2,
-            "Should produce 2 separate beams (one per measure, not crossing barline)"
-        );
-    }
-
-    /// T053: Beams break at rests
-    #[test]
-    fn test_beams_break_at_rests() {
-        // 2 eighths, then a quarter rest, then 2 more eighths → 2 beam groups
-        let score = serde_json::json!({
-            "instruments": [{
-                "id": "piano",
-                "staves": [{
-                    "clef": "Treble",
-                    "time_signature": { "numerator": 4, "denominator": 4 },
-                    "key_signature": { "sharps": 0 },
-                    "voices": [{
-                        "notes": [
-                            { "pitch": 72, "tick": 0, "duration": 480 },
-                            { "pitch": 74, "tick": 480, "duration": 480 },
-                            { "pitch": 76, "tick": 1920, "duration": 480 },
-                            { "pitch": 77, "tick": 2400, "duration": 480 }
-                        ]
-                    }]
-                }]
-            }]
-        });
-
-        let config = LayoutConfig::default();
-        let layout = compute_layout(&score, &config);
-
-        let staff = &layout.systems[0].staff_groups[0].staves[0];
-        let all_glyphs: Vec<_> = staff
-            .glyph_runs
-            .iter()
-            .flat_map(|run| run.glyphs.iter())
-            .collect();
-
-        let beams = all_glyphs
-            .iter()
-            .filter(|g| g.codepoint == "\u{0001}")
-            .count();
-
-        assert_eq!(
-            beams, 2,
-            "Should produce 2 beams (broken by rest gap between beats)"
-        );
-    }
-
-    /// Test compute_staff_note_extents returns staff-line bounds when no notes extend beyond
-    #[test]
-    fn test_compute_staff_note_extents_within_staff() {
-        let ups = 20.0;
-        let staff_data = StaffData {
-            clef: "Treble".to_string(),
-            time_numerator: 4,
-            time_denominator: 4,
-            key_sharps: 0,
-            key_signature_events: vec![],
-            clef_events: vec![],
-            voices: vec![VoiceData {
-                notes: vec![NoteEvent {
-                    pitch: 67, // G4 — on the second line of treble staff
-                    start_tick: 0,
-                    duration_ticks: 960,
-                    spelling: None,
-                    beam_info: vec![],
-                    staccato: false,
-                    dot_count: 0,
-                    note_id: String::new(),
-                    tie_next: None,
-                    slur_next: None,
-                    slur_above: None,
-                }],
-                rests: vec![],
-            }],
-        };
-        let tick_range = TickRange {
-            start_tick: 0,
-            end_tick: 3840,
-        };
-        let (min_y, max_y) = compute_staff_note_extents(&staff_data, &tick_range, ups);
-        // G4 anchor y=50 (below middle line at 40), stem goes UP: tip = 50-70 = -20.
-        // Stem-aware extents include this above-staff stem.
-        assert!(
-            (min_y - (-20.0)).abs() < 0.1,
-            "min_y should be ~-20 for G4 (stem extends above staff), got {}",
-            min_y
-        );
-        assert!(
-            (max_y - 4.0 * ups).abs() < 0.1,
-            "max_y should be ~4*ups for in-staff note, got {}",
-            max_y
-        );
-    }
-
-    /// Test compute_staff_note_extents extends downward for low notes
-    #[test]
-    fn test_compute_staff_note_extents_below_staff() {
-        let ups = 20.0;
-        let staff_data = StaffData {
-            clef: "Treble".to_string(),
-            time_numerator: 4,
-            time_denominator: 4,
-            key_sharps: 0,
-            key_signature_events: vec![],
-            clef_events: vec![],
-            voices: vec![VoiceData {
-                notes: vec![NoteEvent {
-                    pitch: 60, // C4 (middle C) — below treble staff, needs ledger line
-                    start_tick: 0,
-                    duration_ticks: 960,
-                    spelling: None,
-                    beam_info: vec![],
-                    staccato: false,
-                    dot_count: 0,
-                    note_id: String::new(),
-                    tie_next: None,
-                    slur_next: None,
-                    slur_above: None,
-                }],
-                rests: vec![],
-            }],
-        };
-        let tick_range = TickRange {
-            start_tick: 0,
-            end_tick: 3840,
-        };
-        let (_min_y, max_y) = compute_staff_note_extents(&staff_data, &tick_range, ups);
-        // C4 in treble clef is below the staff (y > 4*ups)
-        assert!(
-            max_y > 4.0 * ups,
-            "max_y should extend below staff for C4 in treble, got {}",
-            max_y
         );
     }
