+++ specs/052-refactor-layout-engine/checklists/requirements.md
@@ -1,0 +1,36 @@
+# Specification Quality Checklist: Refactor Layout Engine
+
+**Purpose**: Validate specification completeness and quality before proceeding to planning
+**Created**: 2025-03-18
+**Feature**: [spec.md](../spec.md)
+
+## Content Quality
+
+- [x] No implementation details (languages, frameworks, APIs)
+- [x] Focused on user value and business needs
+- [x] Written for non-technical stakeholders
+- [x] All mandatory sections completed
+
+## Requirement Completeness
+
+- [x] No [NEEDS CLARIFICATION] markers remain
+- [x] Requirements are testable and unambiguous
+- [x] Success criteria are measurable
+- [x] Success criteria are technology-agnostic (no implementation details)
+- [x] All acceptance scenarios are defined
+- [x] Edge cases are identified
+- [x] Scope is clearly bounded
+- [x] Dependencies and assumptions identified
+
+## Feature Readiness
+
+- [x] All functional requirements have clear acceptance criteria
+- [x] User scenarios cover primary flows
+- [x] Feature meets measurable outcomes defined in Success Criteria
+- [x] No implementation details leak into specification
+
+## Notes
+
+- All items pass validation. The spec references file names (`mod.rs`, `types.rs`, etc.) as domain entities rather than implementation choices — these are the subject of the refactoring, not prescriptions of how to implement.
+- The spec is intentionally scoped as a pure structural refactoring with zero behavioral changes.
+- Ready to proceed to `/speckit.clarify` or `/speckit.plan`.
